test(replication): split replication_test into logical suites
Split the 5000-line replication_test.py into four focused files that
share helpers from replication_utils.py, keeping the original module as
the "core" suite.

- replication_test.py: full/stable sync, scripts, expiry, info, role, wait.
- replication_resilience_test.py: disconnect/reconnect/crash, takeover,
  timeouts, partial sync.
- replication_config_test.py: TLS/admin/ipv6/announce, replicaof flags,
  auth/ACL, old-master compat, mismatched dbnum.
- replication_specific_test.py: big values/snapshot/streaming, eviction,
  memory, search/HNSW, streams/memcached, SBF/CF chunked.
- Move cross-suite helpers get_metric_value and wait_for_replica_status
  into replication_utils.py.

All 94 tests preserved (149/193 collected, unchanged).
