+++ tests/dragonfly/replication_test.py
@@ -1,30 +1,22 @@
 import asyncio
 import logging
-import os
-import platform
 import random
 import re
-import shutil
-import signal
-import struct
-import tarfile
 import time
-from itertools import chain, repeat
 
 import async_timeout
-import pymemcache
 import pytest
 import redis
 from redis import asyncio as aioredis
 
 from . import dfly_args
-from .instance import DflyInstanceFactory, DflyInstance
-from .seeder import DebugPopulateSeeder, HnswSearchSeeder
+from .instance import DflyInstanceFactory
+from .seeder import DebugPopulateSeeder
 from .seeder import Seeder as SeederV2
 from .replication_utils import (
     ADMIN_PORT,
-    assert_debug_populate_synced,
     assert_replica_data_matches,
+    get_metric_value,
     master_role_reply,
     replica_role_reply,
     setup_replication,
@@ -34,15 +26,8 @@ from .utility import (
     assert_eventually,
     batch_fill_data,
     check_all_replicas_finished,
-    DflySeederFactory,
-    download_with_retries,
-    ExpirySeeder,
-    extract_int_after_prefix,
     gen_test_data,
-    info_tick_timer,
-    is_saving,
     parse_client_list,
-    tmp_file_name,
     wait_available_async,
     wait_for_replicas_state,
 )
@@ -53,6 +38,8 @@ DISCONNECT_NORMAL_STABLE_SYNC = 2
 
 M_OPT = [pytest.mark.opt_only]
 M_SLOW = [pytest.mark.large]
+
+
 M_STRESS = [pytest.mark.large, pytest.mark.opt_only]
 M_NOT_EPOLL = [pytest.mark.exclude_epoll]
 
@@ -242,321 +229,6 @@ async def test_replication_list_double_apply(df_factory: DflyInstanceFactory):
     await stream_task
 
     await assert_replica_data_matches(c_master, [c_replica])
-
-
-"""
-Test disconnecting replicas during different phases while constantly streaming changes to master.
-
-This test is targeted at the master cancellation mechanism that should qickly stop operations for a
-disconnected replica.
-
-Three types are tested:
-1. Replicas crashing during full sync state
-2. Replicas crashing during stable sync state
-3. Replicas disconnecting normally with REPLICAOF NO ONE during stable state
-"""
-
-# 1. Number of master threads
-# 2. Number of threads for each replica that crashes during full sync
-# 3. Number of threads for each replica that crashes during stable sync
-# 4. Number of threads for each replica that disconnects normally
-# 5. Number of distinct keys that are constantly streamed
-disconnect_cases = [
-    # balanced
-    (8, [4, 4], [4, 4], [4], 4_000),
-    (4, [2] * 4, [2] * 4, [2, 2], 2_000),
-    # full sync heavy
-    (8, [4] * 4, [], [], 4_000),
-    # stable state heavy
-    (8, [], [4] * 4, [], 4_000),
-    # disconnect only
-    (8, [], [], [4] * 4, 4_000),
-]
-
-
-@pytest.mark.parametrize("t_master, t_crash_fs, t_crash_ss, t_disonnect, n_keys", disconnect_cases)
-async def test_disconnect_replica(
-    df_factory: DflyInstanceFactory,
-    df_seeder_factory,
-    t_master,
-    t_crash_fs,
-    t_crash_ss,
-    t_disonnect,
-    n_keys,
-):
-    master = df_factory.create(
-        proactor_threads=t_master, vmodule="replica=2,dflycmd=2,server_family=2"
-    )
-    replicas = [
-        (
-            df_factory.create(proactor_threads=t, vmodule="replica=2,dflycmd=2,server_family=2"),
-            crash_fs,
-        )
-        for i, (t, crash_fs) in enumerate(
-            chain(
-                zip(t_crash_fs, repeat(DISCONNECT_CRASH_FULL_SYNC)),
-                zip(t_crash_ss, repeat(DISCONNECT_CRASH_STABLE_SYNC)),
-                zip(t_disonnect, repeat(DISCONNECT_NORMAL_STABLE_SYNC)),
-            )
-        )
-    ]
-
-    logging.debug("Start master")
-    master.start()
-    c_master = master.client(single_connection_client=True)
-
-    logging.debug("Start replicas and create clients")
-    df_factory.start_all([replica for replica, _ in replicas])
-
-    c_replicas = [(replica, replica.client(), crash_type) for replica, crash_type in replicas]
-
-    def replicas_of_type(tfunc):
-        return [args for args in c_replicas if tfunc(args[2])]
-
-    logging.debug("Start data fill loop")
-    seeder = df_seeder_factory.create(port=master.port, keys=n_keys, dbcount=2)
-    fill_task = asyncio.create_task(seeder.run())
-
-    logging.debug("Run full sync")
-
-    async def full_sync(replica: DflyInstance, c_replica, crash_type):
-        await c_replica.execute_command("REPLICAOF localhost " + str(master.port))
-        if crash_type == 0:
-            await asyncio.sleep(random.random() / 100 + 0.01)
-            await c_replica.aclose()
-            replica.stop(kill=True)
-        else:
-            await wait_available_async(c_replica)
-
-    await asyncio.gather(*(full_sync(*args) for args in c_replicas))
-
-    # Wait for master to stream a bit more
-    await asyncio.sleep(0.1)
-
-    # Check master survived full sync crashes
-    assert await c_master.ping()
-
-    # Check phase-2 replicas survived
-    for _, c_replica, _ in replicas_of_type(lambda t: t > 0):
-        assert await c_replica.ping()
-
-    logging.debug("Run stable state crashes")
-
-    async def stable_sync(replica, c_replica, crash_type):
-        await asyncio.sleep(random.random() / 100)
-        await c_replica.aclose()
-        replica.stop(kill=True)
-
-    await asyncio.gather(*(stable_sync(*args) for args in replicas_of_type(lambda t: t == 1)))
-
-    # Check master survived all crashes
-    assert await c_master.ping()
-
-    # Check phase 3 replica survived
-    for _, c_replica, _ in replicas_of_type(lambda t: t > 1):
-        assert await c_replica.ping()
-
-    logging.debug("Check master survived all crashes")
-    assert await c_master.ping()
-
-    # Check disconnects
-    async def disconnect(replica, c_replica, crash_type):
-        await asyncio.sleep(random.random() / 100)
-        await c_replica.execute_command("REPLICAOF NO ONE")
-
-    logging.debug("disconnect replicas")
-    await asyncio.gather(*(disconnect(*args) for args in replicas_of_type(lambda t: t == 2)))
-
-    await asyncio.sleep(0.5)
-
-    logging.debug("Check phase 3 replica survived")
-    for replica, c_replica, _ in replicas_of_type(lambda t: t == 2):
-        assert await c_replica.ping()
-        await c_replica.aclose()
-
-    logging.debug("Stop streaming")
-    seeder.stop()
-    await fill_task
-
-    logging.debug("Check master survived all disconnects")
-    assert await c_master.ping()
-
-
-"""
-Test stopping master during different phases.
-
-This test is targeted at the replica cancellation mechanism that should quickly abort a failed operation
-and revert to connection retry state.
-
-Three types are tested:
-1. Master crashing during full sync state
-2. Master crashing in a random state.
-3. Master crashing during stable sync state
-
-"""
-
-# 1. Number of master threads
-# 2. Number of threads for each replica
-# 3. Number of times a random crash happens
-# 4. Number of keys transferred (the more, the higher the propability to not miss full sync)
-master_crash_cases = [
-    (6, [6], 3, 2_000),
-    (4, [4, 4, 4], 3, 2_000),
-]
-
-
-@pytest.mark.large
-@pytest.mark.parametrize("t_master, t_replicas, n_random_crashes, n_keys", master_crash_cases)
-async def test_disconnect_master(
-    df_factory, df_seeder_factory, t_master, t_replicas, n_random_crashes, n_keys
-):
-    master = df_factory.create(port=1111, proactor_threads=t_master)
-    replicas = [df_factory.create(proactor_threads=t) for i, t in enumerate(t_replicas)]
-
-    df_factory.start_all(replicas)
-    c_replicas = [replica.client() for replica in replicas]
-
-    seeder = df_seeder_factory.create(port=master.port, keys=n_keys, dbcount=2)
-
-    async def crash_master_fs():
-        await asyncio.sleep(random.random() / 10)
-        master.stop(kill=True)
-
-    async def start_master():
-        await asyncio.sleep(0.2)
-        master.start()
-        async with master.client() as c_master:
-            assert await c_master.ping()
-            seeder.reset()
-            await seeder.run(target_deviation=0.1)
-
-    await start_master()
-
-    # Crash master during full sync, but with all passing initial connection phase
-    await asyncio.gather(
-        *(
-            c_replica.execute_command("REPLICAOF localhost " + str(master.port))
-            for c_replica in c_replicas
-        )
-    )
-    await crash_master_fs()
-
-    await asyncio.sleep(1 + len(replicas) * 0.5)
-
-    for _ in range(n_random_crashes):
-        await start_master()
-        await asyncio.sleep(random.random() + len(replicas) * random.random() / 10)
-        # Crash master in some random state for each replica
-        master.stop(kill=True)
-
-    await start_master()
-    await asyncio.sleep(1 + len(replicas) * 0.5)  # Replicas check every 500ms.
-    capture = await seeder.capture()
-    for replica, c_replica in zip(replicas, c_replicas):
-        await wait_available_async(c_replica)
-        assert await seeder.compare(capture, port=replica.port)
-
-    # Crash master during stable state
-    master.stop(kill=True)
-
-    await start_master()
-    await asyncio.sleep(1 + len(replicas) * 0.5)
-    capture = await seeder.capture()
-    for c_replica in c_replicas:
-        await wait_available_async(c_replica)
-        assert await seeder.compare(capture, port=replica.port)
-
-
-"""
-Test re-connecting replica to different masters.
-"""
-
-rotating_master_cases = [(4, [4, 4, 4, 4], dict(keys=2_000, dbcount=4))]
-
-
-@pytest.mark.large
-@pytest.mark.parametrize("t_replica, t_masters, seeder_config", rotating_master_cases)
-async def test_rotating_masters(df_factory, df_seeder_factory, t_replica, t_masters, seeder_config):
-    replica = df_factory.create(proactor_threads=t_replica)
-    masters = [df_factory.create(proactor_threads=t) for i, t in enumerate(t_masters)]
-    df_factory.start_all([replica] + masters)
-
-    seeders = [df_seeder_factory.create(port=m.port, **seeder_config) for m in masters]
-
-    c_replica = replica.client()
-
-    await asyncio.gather(*(seeder.run(target_deviation=0.1) for seeder in seeders))
-
-    fill_seeder = None
-    fill_task = None
-
-    for master, seeder in zip(masters, seeders):
-        if fill_task is not None:
-            fill_seeder.stop()
-            fill_task.cancel()
-
-        await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-        await wait_available_async(c_replica)
-
-        capture = await seeder.capture()
-        assert await seeder.compare(capture, port=replica.port)
-
-        fill_task = asyncio.create_task(seeder.run())
-        fill_seeder = seeder
-
-    if fill_task is not None:
-        fill_seeder.stop()
-        fill_task.cancel()
-
-
-@pytest.mark.large
-async def test_cancel_replication_immediately(df_factory, df_seeder_factory: DflySeederFactory):
-    """
-    Issue 100 replication commands. This checks that the replication state
-    machine can handle cancellation well.
-    """
-    COMMANDS_TO_ISSUE = 100
-
-    replica = df_factory.create()
-    master = df_factory.create()
-    df_factory.start_all([replica, master])
-
-    seeder = df_seeder_factory.create(port=master.port)
-    c_replica = replica.client(socket_timeout=80)
-
-    await seeder.run(target_deviation=0.1)
-
-    async def ping_status():
-        while True:
-            await c_replica.info()
-            await asyncio.sleep(0.05)
-
-    async def replicate():
-        await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-        return True
-
-    ping_job = asyncio.create_task(ping_status())
-    replication_commands = [asyncio.create_task(replicate()) for _ in range(COMMANDS_TO_ISSUE)]
-
-    num_successes = 0
-    for result in asyncio.as_completed(replication_commands, timeout=80):
-        num_successes += await result
-
-    logging.info(f"succeses: {num_successes}")
-    assert COMMANDS_TO_ISSUE == num_successes
-
-    await wait_available_async(c_replica)
-    capture = await seeder.capture()
-    logging.info(f"number of items captured {len(capture)}")
-    assert await seeder.compare(capture, replica.port)
-
-    ping_job.cancel()
-
-    replica.stop()
-    lines = replica.find_in_logs("Stopping replication")
-    # Cancelled 99 times by REPLICAOF command and once by Shutdown() because
-    # we stopped the instance
-    assert len(lines) == COMMANDS_TO_ISSUE
 
 
 """
@@ -1078,35 +750,6 @@ async def test_scripts(df_factory, t_master, t_replicas, num_ops, num_keys, num_
                 assert l == [f"{j}"] * num_ops
 
 
-@dfly_args({"proactor_threads": 4})
-async def test_auth_master(df_factory, n_keys=20):
-    masterpass = "requirepass"
-    replicapass = "replicapass"
-    master = df_factory.create(requirepass=masterpass)
-    replica = df_factory.create(masterauth=masterpass, requirepass=replicapass)
-
-    df_factory.start_all([master, replica])
-
-    c_master = master.client(password=masterpass)
-    c_replica = replica.client(password=replicapass)
-
-    # Connect replica to master
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-
-    # Set keys
-    pipe = c_master.pipeline(transaction=False)
-    batch_fill_data(pipe, gen_test_data(n_keys))
-    await pipe.execute()
-
-    # Check replica finished executing the replicated commands
-    await check_all_replicas_finished([c_replica], c_master)
-    # Check keys are on replica
-    res = await c_replica.mget(k for k, _ in gen_test_data(n_keys))
-    assert all(v is not None for v in res)
-    await c_master.connection_pool.disconnect()
-    await c_replica.connection_pool.disconnect()
-
-
 SCRIPT_TEMPLATE = "return {}"
 
 
@@ -1178,10 +821,6 @@ def parse_lag(replication_info: str):
     return int(lags[0])
 
 
-async def get_metric_value(inst, metric_name, sample_index=0):
-    return (await inst.metrics())[metric_name].samples[sample_index].value
-
-
 async def assert_lag_condition(inst, client, condition):
     """
     Since lag is a bit random, and we want stable tests, we check
@@ -1206,23 +845,6 @@ async def assert_lag_condition(inst, client, condition):
         await asyncio.sleep(0.05)
     else:
         assert False, "Lag has never satisfied condition!"
-
-
-async def get_replica_reconnects_count(replica_inst):
-    return await get_metric_value(replica_inst, "dragonfly_replica_reconnect_count")
-
-
-async def assert_replica_reconnections(replica_inst, initial_reconnects_count):
-    """
-    Asserts that the replica has attempted to reconnect at least once.
-    """
-    reconnects_count = await get_replica_reconnects_count(replica_inst)
-    if reconnects_count > initial_reconnects_count:
-        return
-
-    assert (
-        False
-    ), f"Expected reconnect count to increase by at least 1, but it did not. Initial dragonfly_replica_reconnect_count: {initial_reconnects_count}, current count: {reconnects_count}"
 
 
 @dfly_args({"proactor_threads": 2})
@@ -1342,661 +964,6 @@ async def test_readonly_script(df_factory):
         await c_replica.eval(WRITE_SCRIPT, 1, "A")
 
 
-take_over_cases = [
-    [2, 2],
-    [2, 4],
-    [4, 2],
-    [8, 8],
-]
-
-
-@pytest.mark.exclude_epoll
-@pytest.mark.parametrize("master_threads, replica_threads", take_over_cases)
-async def test_take_over_counters(df_factory, master_threads, replica_threads):
-    master = df_factory.create(proactor_threads=master_threads)
-    replica1 = df_factory.create(proactor_threads=replica_threads)
-    replica2 = df_factory.create(proactor_threads=replica_threads)
-    replica3 = df_factory.create(proactor_threads=replica_threads)
-    df_factory.start_all([master, replica1, replica2, replica3])
-    c_master = master.client()
-    c1 = replica1.client()
-    c_blocking = master.client()
-    c2 = replica2.client()
-    c3 = replica3.client()
-
-    await c1.execute_command(f"REPLICAOF localhost {master.port}")
-    await c2.execute_command(f"REPLICAOF localhost {master.port}")
-    await c3.execute_command(f"REPLICAOF localhost {master.port}")
-
-    await wait_available_async(c1)
-
-    async def counter(key):
-        value = 0
-        await c_master.execute_command(f"SET {key} 0")
-        start = time.time()
-        while time.time() - start < 20:
-            try:
-                value = await c_master.execute_command(f"INCR {key}")
-            except (redis.exceptions.ConnectionError, redis.exceptions.ResponseError):
-                break
-        else:
-            assert False, "The incrementing loop should be exited with a connection error"
-        return key, value
-
-    async def block_during_takeover():
-        "Add a blocking command during takeover to make sure it doesn't block it."
-        start = time.time()
-        # The command should just be canceled
-        assert await c_blocking.execute_command("BLPOP BLOCKING_KEY1 BLOCKING_KEY2 100") is None
-        # And it should happen in reasonable amount of time.
-        assert time.time() - start < 10
-
-    async def delayed_takeover():
-        await asyncio.sleep(1)
-        await c1.execute_command("REPLTAKEOVER 5")
-
-    _, _, *results = await asyncio.gather(
-        delayed_takeover(), block_during_takeover(), *[counter(f"key{i}") for i in range(16)]
-    )
-    assert await c1.execute_command("role") == master_role_reply([])
-
-    for key, client_value in results:
-        replicated_value = await c1.get(key)
-        assert client_value == int(replicated_value)
-
-
-@pytest.mark.exclude_epoll
-@pytest.mark.parametrize("master_threads, replica_threads", take_over_cases)
-async def test_take_over_seeder(
-    request, df_factory, df_seeder_factory, master_threads, replica_threads
-):
-    master = df_factory.create(
-        proactor_threads=master_threads, dbfilename=f"dump_{tmp_file_name()}", admin_port=ADMIN_PORT
-    )
-    replica = df_factory.create(proactor_threads=replica_threads)
-    df_factory.start_all([master, replica])
-
-    seeder = df_seeder_factory.create(port=master.port, keys=1000, dbcount=5, stop_on_failure=False)
-
-    c_replica = replica.client()
-
-    await start_replication(c_replica, master.admin_port)
-
-    fill_task = asyncio.create_task(seeder.run())
-
-    stop_info = False
-
-    async def info_replication():
-        my_client = replica.client()
-        while not stop_info:
-            await my_client.info("replication")
-            await asyncio.sleep(0.5)
-
-    info_task = asyncio.create_task(info_replication())
-
-    # Give the seeder a bit of time.
-    await asyncio.sleep(3)
-    logging.debug("running repltakover")
-    await c_replica.execute_command("REPLTAKEOVER 30 SAVE")
-    logging.debug("after running repltakover")
-    seeder.stop()
-    await fill_task
-
-    assert await c_replica.execute_command("role") == master_role_reply([])
-    stop_info = True
-    await info_task
-
-    @assert_eventually
-    async def assert_master_exists():
-        assert master.proc.poll() == 0, "Master process did not exit correctly."
-
-    await assert_master_exists()
-
-    master.start()
-    c_master = master.client()
-    await wait_available_async(c_master)
-
-    capture = await seeder.capture(port=master.port)
-    assert await seeder.compare(capture, port=replica.port)
-
-
-@pytest.mark.parametrize("master_threads, replica_threads", [[4, 4]])
-async def test_take_over_read_commands(df_factory, master_threads, replica_threads):
-    master = df_factory.create(proactor_threads=master_threads)
-    replica = df_factory.create(proactor_threads=replica_threads)
-    df_factory.start_all([master, replica])
-
-    c_master = master.client(socket_timeout=1, socket_connect_timeout=1)
-    await c_master.execute_command("SET foo bar")
-
-    c_replica = replica.client()
-    await start_replication(c_replica, master.port)
-
-    async def prompt():
-        client = replica.client()
-        master_alive = True
-        for i in range(10):
-            # TODO remove try block when we no longer shut down master after take over
-            if master_alive:
-                try:
-                    res = await c_master.execute_command("GET foo")
-                    assert res == "bar"
-                    res = await c_master.execute_command("CONFIG SET aclfile myfile")
-                    assert res == "OK"
-                except:
-                    master_alive = False
-            res = await client.execute_command("GET foo")
-            assert res == "bar"
-
-    promt_task = asyncio.create_task(prompt())
-    await c_replica.execute_command("REPLTAKEOVER 5")
-
-    assert await c_replica.execute_command("role") == master_role_reply([])
-    await promt_task
-
-
-async def test_take_over_timeout(df_factory, df_seeder_factory):
-    master = df_factory.create(proactor_threads=2)
-    replica = df_factory.create(proactor_threads=2)
-    df_factory.start_all([master, replica])
-
-    seeder = df_seeder_factory.create(port=master.port, keys=1000, dbcount=5, stop_on_failure=False)
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    logging.debug(f"PORTS ARE:  {master.port} {replica.port}")
-
-    await start_replication(c_replica, master.port)
-
-    fill_task = asyncio.create_task(seeder.run(target_ops=3000))
-
-    # Give the seeder a bit of time.
-    await asyncio.sleep(1)
-    try:
-        await c_replica.execute_command("REPLTAKEOVER 0")
-    except redis.exceptions.ResponseError as e:
-        # Should fail with detailed error message
-        assert str(e).startswith("Couldn't execute takeover")
-        # Verify it includes diagnostic information
-        assert ":" in str(e), "Error message should include diagnostic details"
-    else:
-        assert False, "Takeover should not succeed."
-    seeder.stop()
-    await fill_task
-
-    assert await c_master.execute_command("role") == master_role_reply(replica)
-    assert await c_replica.execute_command("role") == replica_role_reply(master)
-
-
-# 1. Number of master threads
-# 2. Number of threads for each replica
-replication_cases = [(8, 8)]
-
-
-@pytest.mark.parametrize("t_master, t_replica", replication_cases)
-async def test_no_tls_on_admin_port(
-    df_factory: DflyInstanceFactory,
-    df_seeder_factory,
-    t_master,
-    t_replica,
-    with_tls_server_args,
-):
-    # 1. Spin up dragonfly without tls, debug populate
-    master = df_factory.create(
-        no_tls_on_admin_port="true",
-        admin_port=ADMIN_PORT,
-        **with_tls_server_args,
-        requirepass="XXX",
-        proactor_threads=t_master,
-    )
-    master.start()
-    c_master = master.admin_client(password="XXX")
-    await c_master.execute_command("DEBUG POPULATE 100")
-    db_size = await c_master.execute_command("DBSIZE")
-    assert 100 == db_size
-
-    # 2. Spin up a replica and initiate a REPLICAOF
-    replica = df_factory.create(
-        no_tls_on_admin_port="true",
-        admin_port=ADMIN_PORT + 1,
-        **with_tls_server_args,
-        proactor_threads=t_replica,
-        requirepass="XXX",
-        masterauth="XXX",
-    )
-    replica.start()
-    c_replica = replica.admin_client(password="XXX")
-    res = await c_replica.execute_command("REPLICAOF localhost " + str(master.admin_port))
-    assert "OK" == res
-
-    # 3. Verify that replica dbsize == debug populate key size -- replication works
-    await assert_debug_populate_synced(c_master, c_replica, 100, populate=False)
-
-
-# 1. Number of master threads
-# 2. Number of threads for each replica
-# 3. Admin port
-replication_cases = [(8, 8, False), (8, 8, True)]
-
-
-@pytest.mark.parametrize("t_master, t_replica, test_admin_port", replication_cases)
-async def test_tls_replication(
-    df_factory,
-    df_seeder_factory,
-    t_master,
-    t_replica,
-    test_admin_port,
-    with_ca_tls_server_args,
-    with_ca_tls_client_args,
-    proxy_factory,
-):
-    # 1. Spin up dragonfly tls enabled, debug populate
-    master = df_factory.create(
-        tls_replication="true",
-        **with_ca_tls_server_args,
-        port=1111,
-        admin_port=ADMIN_PORT,
-        proactor_threads=t_master,
-    )
-    master.start()
-    c_master = master.client(**with_ca_tls_client_args)
-    await c_master.execute_command("DEBUG POPULATE 100")
-    db_size = await c_master.execute_command("DBSIZE")
-    assert 100 == db_size
-
-    proxy = await proxy_factory(master.port if not test_admin_port else master.admin_port)
-
-    # 2. Spin up a replica and initiate a REPLICAOF
-    replica = df_factory.create(
-        tls_replication="true",
-        **with_ca_tls_server_args,
-        proactor_threads=t_replica,
-    )
-    replica.start()
-    c_replica = replica.client(**with_ca_tls_client_args)
-    res = await c_replica.execute_command("REPLICAOF localhost " + str(proxy.port))
-    assert "OK" == res
-
-    # 3. Verify that replica dbsize == debug populate key size -- replication works
-    await assert_debug_populate_synced(c_master, c_replica, 100, populate=False)
-
-    # 4. Break the connection between master and replica
-    await proxy.close()
-    await asyncio.sleep(3)
-    await proxy.start_serving()
-
-    # Check replica gets new keys
-    await c_master.execute_command("SET MY_KEY 1")
-    db_size = await c_master.execute_command("DBSIZE")
-    assert 101 == db_size
-
-    await check_all_replicas_finished([c_replica], c_master)
-    db_size = await c_replica.execute_command("DBSIZE")
-    assert 101 == db_size
-
-
-@dfly_args({"proactor_threads": 2})
-async def test_tls_replication_without_ca(
-    df_factory,
-    df_seeder_factory,
-    with_tls_server_args,
-    with_ca_tls_client_args,
-):
-    # 1. Spin up dragonfly tls enabled, debug populate
-    master = df_factory.create(tls_replication="true", **with_tls_server_args, requirepass="hi")
-    master.start()
-    # Somehow redis-py forces to verify the certificate and it fails
-    # TODO investigate why and remove with_ca_tls_clients_args
-    c_master = master.client(password="hi", **with_ca_tls_client_args)
-    await c_master.execute_command("DEBUG POPULATE 100")
-
-    # 2. Spin up a replica and initiate a REPLICAOF
-    replica = df_factory.create(
-        tls_replication="true", **with_tls_server_args, masterauth="hi", requirepass="hi"
-    )
-    replica.start()
-
-    c_replica = replica.client(password="hi", **with_ca_tls_client_args)
-
-    res = await c_replica.execute_command("REPLICAOF localhost " + str(master.port))
-    assert "OK" == res
-    await assert_debug_populate_synced(c_master, c_replica, 100, populate=False)
-
-
-@pytest.mark.exclude_epoll
-async def test_ipv6_replication(df_factory: DflyInstanceFactory):
-    """Test that IPV6 addresses work for replication, ::1 is 127.0.0.1 localhost"""
-    master = df_factory.create(proactor_threads=1, bind="::1", port=1111)
-    replica = df_factory.create(proactor_threads=1, bind="::1", port=1112)
-
-    df_factory.start_all([master, replica])
-    c_master = master.client()
-    c_replica = replica.client()
-
-    assert await c_master.ping()
-    assert await c_replica.ping()
-    assert await c_replica.execute_command("REPLICAOF", master["bind"], master["port"]) == "OK"
-
-
-# busy wait for 'replica' instance to have replication status 'status'
-async def wait_for_replica_status(
-    replica: aioredis.Redis, status: str, wait_for_seconds=0.01, timeout=20
-):
-    start = time.time()
-    while (time.time() - start) < timeout:
-        await asyncio.sleep(wait_for_seconds)
-
-        info = await replica.info("replication")
-        if info["master_link_status"] == status:
-            return
-    raise RuntimeError("Client did not become available in time!")
-
-
-@pytest.mark.parametrize(
-    "start_replica_first, disconnect_at_end",
-    [
-        (False, False),
-        (True, False),
-        (False, True),
-    ],
-)
-async def test_replicaof_flag(df_factory, start_replica_first, disconnect_at_end):
-    if start_replica_first:
-        BASE_PORT = 1111
-        # Replica starts before master, uses a fixed port so it knows where to connect
-        replica = df_factory.create(
-            proactor_threads=2,
-            replicaof=f"localhost:{BASE_PORT}",
-        )
-        replica.start()
-        c_replica = replica.client()
-        await wait_for_replica_status(c_replica, status="down")
-
-        # Check that it is in replica mode but status is down
-        info = await c_replica.info("replication")
-        assert info["role"] == "slave"
-        assert info["master_host"] == "localhost"
-        assert info["master_port"] == BASE_PORT
-        assert info["master_link_status"] == "down"
-
-        master = df_factory.create(
-            port=BASE_PORT,
-            proactor_threads=2,
-        )
-        master.start()
-        c_master = master.client()
-        await c_master.set("KEY", "VALUE")
-        db_size = await c_master.dbsize()
-        assert 1 == db_size
-
-        await wait_for_replica_status(c_replica, status="up")
-        await check_all_replicas_finished([c_replica], c_master)
-    else:
-        master = df_factory.create(proactor_threads=2)
-        master.start()
-        c_master = master.client()
-        if disconnect_at_end:
-            await wait_available_async(c_master)
-        await c_master.set("KEY", "VALUE")
-        db_size = await c_master.dbsize()
-        assert 1 == db_size
-
-        replica = df_factory.create(
-            proactor_threads=2,
-            replicaof=f"localhost:{master.port}",
-        )
-        replica.start()
-        c_replica = replica.client()
-        await wait_available_async(c_replica)
-        await check_all_replicas_finished([c_replica], c_master)
-
-    dbsize = await c_replica.dbsize()
-    assert 1 == dbsize
-
-    val = await c_replica.get("KEY")
-    assert "VALUE" == val
-
-    if disconnect_at_end:
-        await c_replica.replicaof("no", "one")
-        role = await c_replica.role()
-        assert role[0] == "master"
-
-
-async def test_df_crash_on_memcached_error(df_factory):
-    master = df_factory.create(
-        memcached_port=11211,
-        proactor_threads=2,
-    )
-
-    replica = df_factory.create(
-        memcached_port=master.mc_port + 1,
-        proactor_threads=2,
-    )
-
-    master.start()
-    replica.start()
-
-    c_master = master.client()
-    await wait_available_async(c_master)
-
-    c_replica = replica.client()
-    await start_replication(c_replica, master.port)
-
-    memcached_client = pymemcache.Client(f"127.0.0.1:{replica.mc_port}")
-
-    with pytest.raises(pymemcache.exceptions.MemcacheServerError):
-        memcached_client.set("key", "data", noreply=False)
-
-
-async def test_df_crash_on_replicaof_flag(df_factory):
-    master = df_factory.create(
-        proactor_threads=2,
-    )
-    master.start()
-
-    replica = df_factory.create(proactor_threads=2, replicaof=f"127.0.0.1:{master.port}")
-    replica.start()
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    await wait_available_async(c_master)
-    await wait_available_async(c_replica)
-
-    res = await c_replica.execute_command("SAVE DF myfile")
-    assert "OK" == res
-
-    res = await c_replica.execute_command("DBSIZE")
-    assert res == 0
-
-
-@pytest.mark.parametrize(
-    "master_threads, backlog_len, stream_during, stream_target_ops, num_drops, sleep_range, check_stale_log",
-    [
-        (6, None, False, None, 10, (0, 10), False),
-        (4, 4000, True, 4000, 3, (10, 20), False),
-        (4, 1, True, None, 3, (5, 10), True),
-    ],
-)
-async def test_network_disconnect(
-    df_factory,
-    df_seeder_factory,
-    master_threads,
-    backlog_len,
-    stream_during,
-    stream_target_ops,
-    num_drops,
-    sleep_range,
-    check_stale_log,
-    proxy_factory,
-):
-    master_kwargs = dict(proactor_threads=master_threads)
-    if backlog_len is not None:
-        master_kwargs["shard_repl_backlog_len"] = backlog_len
-    master = df_factory.create(**master_kwargs)
-    replica = df_factory.create(proactor_threads=4)
-
-    df_factory.start_all([replica, master])
-    seeder = df_seeder_factory.create(port=master.port)
-
-    async with replica.client() as c_replica, master.client() as c_master:
-        await seeder.run(target_deviation=0.1)
-
-        proxy = await proxy_factory(master.port)
-        await c_replica.execute_command(f"REPLICAOF localhost {proxy.port}")
-
-        if check_stale_log:
-            # Wait for the two nodes to be in sync (stable state replication) before seeding
-            await wait_available_async(c_replica)
-
-        fill_task = None
-        if stream_during:
-            if stream_target_ops is not None:
-                fill_task = asyncio.create_task(seeder.run(target_ops=stream_target_ops))
-            else:
-                fill_task = asyncio.create_task(seeder.run())
-
-        for _ in range(num_drops):
-            await asyncio.sleep(random.randint(*sleep_range) / 10)
-            proxy.drop_connection()
-
-        if fill_task is not None:
-            seeder.stop()
-            await fill_task
-
-        # Give time to detect dropped connection and reconnect
-        await asyncio.sleep(1.0)
-        await wait_available_async(c_replica)
-
-        logging.debug(await c_replica.execute_command("INFO REPLICATION"))
-        logging.debug(await c_master.execute_command("INFO REPLICATION"))
-
-        capture = await seeder.capture()
-        assert await seeder.compare(capture, replica.port)
-
-    if check_stale_log:
-        master.stop()
-        lines = master.find_in_logs("Partial sync requested from stale LSN")
-        assert len(lines) > 0
-
-
-async def test_replica_reconnections_after_network_disconnect(
-    df_factory, df_seeder_factory, proxy_factory
-):
-    master = df_factory.create(proactor_threads=6)
-    replica = df_factory.create(proactor_threads=4)
-
-    df_factory.start_all([replica, master])
-    seeder = df_seeder_factory.create(port=master.port)
-
-    async with replica.client() as c_replica:
-        await seeder.run(target_deviation=0.1)
-
-        proxy = await proxy_factory(master.port)
-        await c_replica.execute_command(f"REPLICAOF localhost {proxy.port}")
-
-        # Wait replica to be up and synchronized with master
-        await wait_available_async(c_replica)
-
-        initial_reconnects_count = await get_replica_reconnects_count(replica)
-
-        # Fully drop the server
-        await proxy.close()
-
-        # After dropping the connection replica should try to reconnect
-        await wait_for_replica_status(c_replica, status="down")
-        await asyncio.sleep(2)
-
-        # Restart the proxy
-        await proxy.start_serving()
-
-        # Wait replica to be reconnected and synchronized with master
-        await wait_available_async(c_replica)
-
-        capture = await seeder.capture()
-        assert await seeder.compare(capture, replica.port)
-
-        # Assert replica reconnects metrics increased
-        await assert_replica_reconnections(replica, initial_reconnects_count)
-
-
-async def test_search(df_factory):
-    master, [replica], c_master, [c_replica] = await setup_replication(
-        df_factory,
-        master_args={"proactor_threads": 4},
-        replica_args={"proactor_threads": 4},
-        connect=False,
-    )
-
-    # First, create an index on replica
-    await c_replica.execute_command("FT.CREATE", "idx-r", "SCHEMA", "f1", "numeric")
-    for i in range(0, 10):
-        await c_replica.hset(f"k{i}", mapping={"f1": i})
-    assert (await c_replica.ft("idx-r").search("@f1:[5 9]")).total == 5
-
-    # Second, create an index on master
-    await c_master.execute_command("FT.CREATE", "idx-m", "SCHEMA", "f2", "numeric")
-    for i in range(0, 10):
-        await c_master.hset(f"k{i}", mapping={"f2": i * 2})
-    assert (await c_master.ft("idx-m").search("@f2:[6 10]")).total == 3
-
-    # Replicate
-    await c_replica.execute_command("REPLICAOF", "localhost", master.port)
-    await wait_available_async(c_replica)
-
-    # Check master index was picked up and original index was deleted
-    assert (await c_replica.execute_command("FT._LIST")) == ["idx-m"]
-
-    # Check query from master runs on replica
-    assert (await c_replica.ft("idx-m").search("@f2:[6 10]")).total == 3
-
-    # Set a new key
-    await c_master.hset("kNEW", mapping={"f2": 100})
-    await asyncio.sleep(0.1)
-
-    assert (await c_replica.ft("idx-m").search("@f2:[100 100]")).docs[0].id == "kNEW"
-
-    # Create a new aux index on master
-    await c_master.execute_command("FT.CREATE", "idx-m2", "SCHEMA", "f2", "numeric", "sortable")
-    await asyncio.sleep(0.1)
-
-    from redis.commands.search.query import Query
-
-    assert (await c_replica.ft("idx-m2").search(Query("*").sort_by("f2").paging(0, 1))).docs[
-        0
-    ].id == "k0"
-
-
-@dfly_args({"proactor_threads": 4})
-async def test_search_with_stream(df_factory: DflyInstanceFactory):
-    master, [replica], c_master, [c_replica] = await setup_replication(df_factory, connect=False)
-
-    # fill master with hsets and create index
-    p = c_master.pipeline(transaction=False)
-    for i in range(10_000):
-        p.hset(f"k{i}", mapping={"name": f"name of {i}"})
-    await p.execute()
-
-    await c_master.execute_command("FT.CREATE i1 SCHEMA name text")
-
-    # start replication and issue one add command and delete commands on master in parallel
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-    await c_master.hset("secret-key", mapping={"name": "new-secret"})
-    for i in range(1_000):
-        await c_master.delete(f"k{i}")
-
-    # expect replica to see only 10k - 1k + 1 = 9001 keys in it's index
-    await wait_available_async(c_replica)
-    await check_all_replicas_finished([c_replica], c_master)
-    assert await c_replica.execute_command("FT.SEARCH i1 * LIMIT 0 0") == [9_001]
-    assert await c_replica.execute_command('FT.SEARCH i1 "secret"') == [
-        1,
-        "secret-key",
-        ["name", "new-secret"],
-    ]
-
-
 # @pytest.mark.large
 @pytest.mark.replication(master_args={"proactor_threads": 4}, replica_args={"proactor_threads": 4})
 async def test_client_pause_with_replica(replication, df_seeder_factory):
@@ -2031,1707 +998,6 @@ async def test_client_pause_with_replica(replication, df_seeder_factory):
 
     capture = await seeder.capture(port=master.port)
     assert await seeder.compare(capture, port=replica.port)
-
-
-@pytest.mark.debug_only
-@dfly_args({"proactor_threads": 2})
-async def test_replicaof_reject_on_load(df_factory, df_seeder_factory):
-    master = df_factory.create()
-    replica = df_factory.create(dbfilename=f"dump_{tmp_file_name()}")
-    df_factory.start_all([master, replica])
-
-    c_replica = replica.client()
-
-    await c_replica.execute_command("DEBUG POPULATE 1000 key 500 RAND type set elements 500")
-
-    replica.stop()
-    replica.start()
-    # Disable retries so that BusyLoadingError is raised immediately.
-    # redis-py >= 7 retries on ConnectionError by default, and BusyLoadingError
-    # inherits from ConnectionError, causing the REPLICAOF to be silently
-    # retried until loading finishes.
-    from redis.retry import Retry
-    from redis.backoff import NoBackoff
-
-    c_replica = replica.client(retry=Retry(NoBackoff(), 0))
-
-    @assert_eventually
-    async def check_replica_isloading():
-        persistence = await c_replica.info("PERSISTENCE")
-        assert persistence["loading"] == 1
-
-    # If this fails adjust load of DEBUG POPULATE above.
-    await check_replica_isloading()
-
-    # Check replica of not alowed while loading snapshot
-    # Keep in mind that if the exception has not been raised, it doesn't mean
-    # that there is a bug because it could be the case that while executing
-    # INFO PERSISTENCE df is in loading state but when we call REPLICAOF df
-    # is no longer in loading state and the assertion false is triggered.
-    with pytest.raises(aioredis.BusyLoadingError):
-        await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-
-    # Check one we finish loading snapshot replicaof success
-    await wait_available_async(c_replica, timeout=180)
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-
-
-async def test_heartbeat_eviction_propagation(df_factory):
-    master, [replica], c_master, [c_replica] = await setup_replication(
-        df_factory,
-        master_args={
-            "proactor_threads": 1,
-            "cache_mode": "true",
-            "maxmemory": "256mb",
-            "enable_heartbeat_eviction": "false",
-        },
-        replica_args={"proactor_threads": 1},
-        connect=False,
-    )
-
-    # fill the master to use about 233mb > 256mb * 0.9, which will trigger heartbeat eviction.
-    await c_master.execute_command("DEBUG POPULATE 233 size 1048576")
-    await start_replication(c_replica, master.port)
-
-    # now enable heart beat eviction
-    await c_master.execute_command("CONFIG SET enable_heartbeat_eviction true")
-
-    while True:
-        info = await c_master.info("stats")
-        evicted_1 = info["evicted_keys"]
-        time.sleep(2)
-        info = await c_master.info("stats")
-        evicted_2 = info["evicted_keys"]
-        if evicted_2 == evicted_1:
-            break
-        else:
-            print("waiting for eviction to finish...", end="\r", flush=True)
-
-    await check_all_replicas_finished([c_replica], c_master)
-    keys_master = await c_master.execute_command("keys *")
-    keys_replica = await c_replica.execute_command("keys *")
-    assert set(keys_master) == set(keys_replica)
-
-
-async def test_policy_based_eviction_propagation(df_factory, df_seeder_factory):
-    master, [replica], c_master, [c_replica] = await setup_replication(
-        df_factory,
-        master_args={
-            "proactor_threads": 2,
-            "cache_mode": "true",
-            "maxmemory": "512mb",
-            "enable_heartbeat_eviction": "false",
-            "rss_oom_deny_ratio": 1.3,
-        },
-        replica_args={"proactor_threads": 2},
-        connect=False,
-    )
-
-    await c_master.execute_command("DEBUG POPULATE 6000 size 88000")
-
-    await start_replication(c_replica, master.port)
-
-    seeder = df_seeder_factory.create(
-        port=master.port, keys=600, val_size=1000, stop_on_failure=False
-    )
-    await seeder.run(target_deviation=0.1)
-
-    info = await c_master.info("stats")
-    assert (
-        info["evicted_keys"] > 0
-    ), f"Weak testcase: policy based eviction was not triggered. {await c_master.info()}"
-
-    await check_all_replicas_finished([c_replica], c_master)
-
-    # KEYS may trigger lazy expiry on master, generating DELs not yet received by replica.
-    # Fetch master keys first, then re-sync to ensure replica applies any resulting DELs.
-    keys_master = await c_master.execute_command("keys k*")
-    await check_all_replicas_finished([c_replica], c_master)
-    keys_replica = await c_replica.execute_command("keys k*")
-
-    assert set(keys_replica).difference(keys_master) == set()
-    assert set(keys_master).difference(keys_replica) == set()
-
-
-async def test_journal_doesnt_yield_issue_2500(df_factory, df_seeder_factory):
-    """
-    Issues many SETEX commands through a Lua script so that no yields are done between them.
-    In parallel, connect a replica, so that these SETEX commands write their custom journal log.
-    This makes sure that no Fiber context switch while inside a shard callback.
-    """
-    master = df_factory.create()
-    replica = df_factory.create()
-    df_factory.start_all([master, replica])
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    async def send_setex():
-        script = """
-        local charset = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890"
-
-        local random_string = function(length)
-            local str = ''
-            for i=1,length do
-                str = str .. charset:sub(math.random(1, #charset))
-            end
-            return str
-        end
-
-        for i = 1, 200 do
-            -- 200 iterations to make sure SliceSnapshot dest queue is full
-            -- 100 bytes string to make sure serializer is big enough
-            redis.call('SETEX', KEYS[1], 1000, random_string(100))
-        end
-        """
-
-        for i in range(10):
-            await asyncio.gather(
-                *[c_master.eval(script, 1, random.randint(0, 1_000)) for j in range(3)]
-            )
-
-    stream_task = asyncio.create_task(send_setex())
-    await asyncio.sleep(0.1)
-
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-    assert not stream_task.done(), "Weak testcase. finished sending commands before replication."
-
-    await wait_available_async(c_replica)
-    await stream_task
-
-    await check_all_replicas_finished([c_replica], c_master)
-    keys_master = await c_master.execute_command("keys *")
-    keys_replica = await c_replica.execute_command("keys *")
-    assert set(keys_master) == set(keys_replica)
-
-
-@pytest.mark.parametrize(
-    "action_during_save",
-    [
-        pytest.param("disconnect", marks=pytest.mark.large),
-        "start_replicating",
-    ],
-)
-async def test_save_with_replication(df_factory, action_during_save):
-    if action_during_save == "disconnect":
-        master = df_factory.create(proactor_threads=1)
-        replica = df_factory.create(proactor_threads=1, dbfilename=f"dump_{tmp_file_name()}")
-    else:
-        master = df_factory.create(proactor_threads=4)
-        replica = df_factory.create(proactor_threads=4, dbfilename=f"dump_{tmp_file_name()}")
-
-    df_factory.start_all([master, replica])
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    if action_during_save == "disconnect":
-        await c_master.execute_command("DEBUG POPULATE 100000 key 4048 RAND")
-        await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-        await wait_available_async(c_replica)
-    else:
-        await c_replica.execute_command("DEBUG POPULATE 100000 key 4096 RAND")
-
-    async def save_replica():
-        await c_replica.execute_command("save")
-
-    save_task = asyncio.create_task(save_replica())
-    while not await is_saving(c_replica):  # wait for replica start saving
-        assert "rdb_changes_since_last_success_save:0" not in await c_replica.execute_command(
-            "info persistence"
-        ), "Weak test case, finished saving too quickly"
-        await asyncio.sleep(0.1)
-
-    if action_during_save == "disconnect":
-        await c_replica.execute_command("replicaof no one")
-    else:
-        await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-
-    assert await is_saving(c_replica)
-    await save_task
-    assert not await is_saving(c_replica)
-
-
-async def test_user_acl_replication(df_factory):
-    master = df_factory.create(proactor_threads=4)
-    replica = df_factory.create(proactor_threads=4)
-    df_factory.start_all([master, replica])
-
-    c_master = master.client()
-    await c_master.execute_command("ACL SETUSER tmp >tmp ON +ping +dfly +replconf")
-    await c_master.execute_command("SET foo bar")
-    assert 1 == await c_master.execute_command("DBSIZE")
-
-    c_replica = replica.client()
-    await c_replica.execute_command("CONFIG SET masteruser tmp")
-    await c_replica.execute_command("CONFIG SET masterauth tmp")
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-
-    await wait_available_async(c_replica)
-    assert 1 == await c_replica.execute_command("DBSIZE")
-
-    # revoke acl's from tmp
-    await c_master.execute_command("ACL SETUSER tmp -replconf")
-    async for info, breaker in info_tick_timer(c_replica, section="REPLICATION"):
-        with breaker:
-            assert info["master_link_status"] == "down"
-
-    await c_master.execute_command("SET bar foo")
-
-    # reinstate and let replication continue
-    await c_master.execute_command("ACL SETUSER tmp +replconf")
-    await check_all_replicas_finished([c_replica], c_master, 5)
-    assert 2 == await c_replica.execute_command("DBSIZE")
-
-
-@pytest.mark.parametrize("break_conn", [False, True])
-async def test_replica_reconnect(df_factory, break_conn):
-    """
-    Test replica does not connect to master if master restarted
-    step1: create master and replica
-    step2: stop master and start again with the same port
-    step3: check replica is not replicating the restarted master
-    step4: issue new replicaof command
-    step5: check replica replicates master
-    """
-    # Connect replica to master
-    master = df_factory.create(proactor_threads=1)
-    replica = df_factory.create(proactor_threads=1, break_replication_on_master_restart=break_conn)
-    df_factory.start_all([master, replica])
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    await c_master.set("k", "12345")
-    await start_replication(c_replica, master.port)
-    assert (await c_replica.info("REPLICATION"))["master_link_status"] == "up"
-
-    # kill existing master, create master with different repl_id but same port
-    master_port = master.port
-    master.stop()
-
-    await asyncio.sleep(1)
-
-    repl_info = await c_replica.info("REPLICATION")
-    assert repl_info["master_link_status"] == "down", str(repl_info)
-
-    master = df_factory.create(proactor_threads=1, port=master_port)
-    df_factory.start_all([master])
-    await asyncio.sleep(1)  # We sleep for 0.5s in replica.cc before reconnecting
-
-    # Assert that replica did not reconnected to master with different repl_id
-    if break_conn:
-        assert await c_master.execute_command("get k") == None
-        assert await c_replica.execute_command("get k") == "12345"
-        assert await c_master.execute_command("set k 6789")
-        assert await c_replica.execute_command("get k") == "12345"
-        assert (await c_replica.info("REPLICATION"))["master_link_status"] == "down"
-    else:
-        assert await c_master.execute_command("get k") == None
-        assert await c_replica.execute_command("get k") == None
-        assert await c_master.execute_command("set k 6789")
-        await check_all_replicas_finished([c_replica], c_master)
-        assert await c_replica.execute_command("get k") == "6789"
-        assert (await c_replica.info("REPLICATION"))["master_link_status"] == "up"
-
-    # Force re-replication, assert that it worked
-    await start_replication(c_replica, master.port)
-    assert await c_replica.execute_command("get k") == "6789"
-
-
-@pytest.mark.replication(
-    replica_args={"replica_announce_ip": "overrode-host", "announce_port": "1337"}
-)
-async def test_announce_ip_port(replication):
-    master, [replica], c_master, [c_replica] = replication
-
-    role, node = await c_master.execute_command("role")
-    assert role == "master"
-    host, port, _ = node[0]
-    assert host == "overrode-host"
-    assert port == "1337"
-
-
-@pytest.mark.parametrize(
-    "with_expiry_seeder",
-    [
-        False,
-        pytest.param(True, marks=pytest.mark.large),
-    ],
-)
-async def test_replication_timeout_on_full_sync(
-    df_factory: DflyInstanceFactory, df_seeder_factory, with_expiry_seeder
-):
-    if with_expiry_seeder:
-        # Timeout set to 3 seconds because we must first saturate the socket such that subsequent
-        # writes block. Otherwise, we will break the flows before Heartbeat actually deadlocks.
-        master = df_factory.create(
-            proactor_threads=2, replication_timeout=3000, vmodule="replica=2,dflycmd=2"
-        )
-    else:
-        # setting replication_timeout to a very small value to force the replica to timeout
-        master = df_factory.create(
-            replication_timeout=100, vmodule="replica=2,dflycmd=2,snapshot=1,rdb_save=1,rdb_load=1"
-        )
-    replica = df_factory.create()
-
-    df_factory.start_all([master, replica])
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    if with_expiry_seeder:
-        await c_master.execute_command("debug", "populate", "100000", "foo", "5000", "RAND")
-
-        c_master = master.client()
-        c_replica = replica.client()
-
-        expiry_seeder = ExpirySeeder()
-        expiry_seeder_task = asyncio.create_task(expiry_seeder.run(c_master))
-        await expiry_seeder.wait_until_n_inserts(50000)
-        expiry_seeder.stop()
-        await expiry_seeder_task
-    else:
-        await c_master.execute_command("debug", "populate", "200000", "foo", "5000", "RAND")
-        seeder = df_seeder_factory.create(port=master.port)
-        seeder_task = asyncio.create_task(seeder.run())
-        await asyncio.sleep(0.5)  # wait for seeder running
-
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-
-    # wait for full sync
-    async with async_timeout.timeout(3):
-        await wait_for_replicas_state(c_replica, state="full_sync", timeout=0.05)
-
-    await c_replica.execute_command(
-        "debug replica pause"
-    )  # pause replica to trigger reconnect on master
-
-    # Dragonfly would get stuck here (with expiry seeder) without the bug fix. When replica does
-    # not read from the socket, Heartbeat() will block on the journal write for the expired items
-    # and shard_handler would never be called and break replication. More details on #3936.
-    pause_duration = 6 if with_expiry_seeder else 1
-    await asyncio.sleep(pause_duration)
-
-    await c_replica.execute_command("debug replica resume")  # resume replication
-
-    await asyncio.sleep(1)  # replica will start resync
-
-    if not with_expiry_seeder:
-        seeder.stop()
-        await seeder_task
-
-    await check_all_replicas_finished([c_replica], c_master, timeout=60)
-    await assert_replica_reconnections(replica, 0)
-
-
-@pytest.mark.exclude_epoll
-@dfly_args({"proactor_threads": 1})
-async def test_master_stalled_disconnect(df_factory: DflyInstanceFactory):
-    # disconnect after 1 second of being blocked
-    master = df_factory.create(replication_timeout=1000)
-    replica = df_factory.create()
-
-    df_factory.start_all([master, replica])
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    await c_master.execute_command("debug", "populate", "200000", "foo", "500", "RAND")
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-
-    @assert_eventually
-    async def check_replica_connected():
-        repl_info = await c_master.info("replication")
-        assert "slave0" in repl_info
-
-    @assert_eventually
-    async def check_replica_disconnected():
-        repl_info = await c_master.info("replication")
-        assert "slave0" not in repl_info
-
-    await check_replica_connected()
-    await c_replica.execute_command("DEBUG REPLICA PAUSE")
-    await check_replica_connected()  # still connected
-    await asyncio.sleep(1)  # wait for the master to recognize it's being blocked
-    await check_replica_disconnected()
-
-
-def download_dragonfly_release(version):
-    arch = platform.machine()  # e.g. "x86_64" or "aarch64"
-    path = f"/tmp/old_df/{version}"
-    binary = f"{path}/dragonfly-{arch}"
-    if os.path.isfile(binary):
-        return binary
-
-    # Cleanup in case there's partial files
-    if os.path.exists(path):
-        shutil.rmtree(path)
-
-    os.makedirs(path)
-    gzfile = f"{path}/dragonfly.tar.gz"
-    url = f"https://github.com/dragonflydb/dragonfly/releases/download/{version}/dragonfly-{arch}.tar.gz"
-    logging.debug(f"Downloading Dragonfly release into {gzfile}...")
-    download_with_retries(url, gzfile)
-
-    # Extract
-    file = tarfile.open(gzfile)
-    file.extractall(path)
-    file.close()
-
-    # Return path
-    return binary
-
-
-@pytest.mark.parametrize(
-    "cluster_mode, announce_ip, announce_port",
-    [
-        ("", "localhost", 7000),
-        ("emulated", "", 0),
-        ("emulated", "localhost", 7000),
-    ],
-)
-async def test_replicate_old_master(
-    df_factory: DflyInstanceFactory, cluster_mode, announce_ip, announce_port
-):
-    cpu = platform.processor()
-    if cpu != "x86_64":
-        pytest.skip(f"Supported only on x64, running on {cpu}")
-
-    dfly_version = "v1.19.2"
-    released_dfly_path = download_dragonfly_release(dfly_version)
-    master = df_factory.create(
-        version=1.19,
-        path=released_dfly_path,
-        cluster_mode=cluster_mode,
-    )
-    replica = df_factory.create(
-        cluster_mode=cluster_mode,
-        cluster_announce_ip=announce_ip,
-        announce_port=announce_port,
-    )
-
-    df_factory.start_all([master, replica])
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    assert (
-        f"df-{dfly_version}"
-        == (await c_master.execute_command("info", "server"))["dragonfly_version"]
-    )
-    assert dfly_version != (await c_replica.execute_command("info", "server"))["dragonfly_version"]
-
-    await c_master.execute_command("set", "k1", "v1")
-
-    assert await c_replica.execute_command(f"REPLICAOF localhost {master.port}") == "OK"
-    await wait_available_async(c_replica)
-
-    assert await c_replica.execute_command("get", "k1") == "v1"
-
-
-# This Test was intorduced in response to a bug when replicating empty hashmaps (encoded as
-# ziplists) created with HSET, HSETEX, HDEL and then replicated 2 times.
-# For more information plz refer to the issue on gh:
-# https://github.com/dragonflydb/dragonfly/issues/3504
-@dfly_args({"proactor_threads": 1})
-async def test_empty_hash_map_replicate_old_master(df_factory):
-    cpu = platform.processor()
-    if cpu != "x86_64":
-        pytest.skip(f"Supported only on x64, running on {cpu}")
-
-    dfly_version = "v1.21.2"
-    released_dfly_path = download_dragonfly_release(dfly_version)
-    # old versions
-    instances = [df_factory.create(path=released_dfly_path, version=1.21) for i in range(3)]
-    # new version
-    instances.append(df_factory.create())
-
-    df_factory.start_all(instances)
-
-    old_c_master = instances[0].client()
-    # Create an empty hashmap
-    await old_c_master.execute_command("HSET foo a_field a_value")
-    await old_c_master.execute_command("HSETEX foo 2 b_field b_value")
-    await old_c_master.execute_command("HDEL foo a_field")
-
-    @assert_eventually
-    async def check_if_empty():
-        assert await old_c_master.execute_command("HGETALL foo") == []
-
-    await check_if_empty()
-    assert await old_c_master.execute_command("EXISTS foo") == 1
-    await old_c_master.aclose()
-
-    async def assert_body(client, result=1, state="online", node_role="slave"):
-        async with async_timeout.timeout(10):
-            await wait_for_replicas_state(client, state=state, node_role=node_role)
-
-        assert await client.execute_command("EXISTS foo") == result
-        assert await client.execute_command("REPLTAKEOVER 1") == "OK"
-
-    index = 0
-    last_old_replica = 2
-
-    # Adjacent pairs
-    for a, b in zip(instances, instances[1:]):
-        logging.debug(index)
-        client_b = b.client()
-        assert await client_b.execute_command(f"REPLICAOF localhost {a.port}") == "OK"
-
-        if index != last_old_replica:
-            await assert_body(client_b, state="stable_sync", node_role="replica")
-        else:
-            await assert_body(client_b, result=0)
-
-        index = index + 1
-        await client_b.aclose()
-
-
-# This Test was intorduced in response to a bug when replicating empty hash maps with
-# HSET, HSETEX, HDEL and then loaded via replication.
-# For more information plz refer to the issue on gh:
-# https://github.com/dragonflydb/dragonfly/issues/3504
-@dfly_args({"proactor_threads": 1})
-async def test_empty_hashmap_loading_bug(df_factory: DflyInstanceFactory):
-    cpu = platform.processor()
-    if cpu != "x86_64":
-        pytest.skip(f"Supported only on x64, running on {cpu}")
-
-    dfly_version = "v1.21.2"
-    released_dfly_path = download_dragonfly_release(dfly_version)
-    master = df_factory.create(path=released_dfly_path, version=1.21)
-    master.start()
-
-    c_master = master.client()
-    # Create an empty hashmap
-    await c_master.execute_command("HSET foo a_field a_value")
-    await c_master.execute_command("HSETEX foo 2 b_field b_value")
-    await c_master.execute_command("HDEL foo a_field")
-
-    @assert_eventually
-    async def check_if_empty():
-        assert await c_master.execute_command("HGETALL foo") == []
-
-    await check_if_empty()
-    assert await c_master.execute_command("EXISTS foo") == 1
-
-    replica = df_factory.create()
-    replica.start()
-    c_replica = replica.client()
-
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-    await wait_for_replicas_state(c_replica)
-    assert await c_replica.execute_command("dbsize") == 0
-
-
-async def test_replicate_search_index_to_old_replica(df_factory: DflyInstanceFactory):
-    """
-    Test that a new master with search indices (including HNSW vector index) can
-    replicate to a v1.35 replica. This verifies backward compatibility of replication
-    when search indices are defined, ensuring the replica receives the data without
-    errors from new RDB AUX fields (search-index, hnsw-index-metadata, HNSW opcodes).
-    """
-    cpu = platform.processor()
-    if cpu != "x86_64":
-        pytest.skip(f"Supported only on x64, running on {cpu}")
-
-    dfly_version = "v1.35.1"
-    released_dfly_path = download_dragonfly_release(dfly_version)
-
-    # New master (current version) with search index
-    master = df_factory.create(proactor_threads=2)
-    # Old replica (v1.35)
-    replica = df_factory.create(
-        version=1.35,
-        path=released_dfly_path,
-        proactor_threads=2,
-    )
-
-    df_factory.start_all([master, replica])
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    # Create a search index with HNSW vector field on the new master
-    await c_master.execute_command(
-        "FT.CREATE",
-        "test_idx",
-        "ON",
-        "HASH",
-        "PREFIX",
-        "1",
-        "item:",
-        "SCHEMA",
-        "name",
-        "TEXT",
-        "price",
-        "NUMERIC",
-        "SORTABLE",
-        "category",
-        "TAG",
-        "embedding",
-        "VECTOR",
-        "HNSW",
-        "6",
-        "TYPE",
-        "FLOAT32",
-        "DIM",
-        "2",
-        "DISTANCE_METRIC",
-        "L2",
-    )
-
-    # Insert test data with vector embeddings
-    for i in range(100):
-        category = "electronics" if i % 2 == 0 else "clothing"
-        embedding = struct.pack("<2f", float(i), float(i * 2))
-        await c_master.hset(
-            f"item:{i}",
-            mapping={
-                "name": f"Product {i}",
-                "price": str(i * 10),
-                "category": category,
-                "embedding": embedding,
-            },
-        )
-
-    # Verify data and index on master
-    assert await c_master.dbsize() == 100
-    master_idx = c_master.ft("test_idx")
-    text_result = await master_idx.search("Product 50")
-    assert text_result.total >= 1
-
-    # Verify KNN search on master. NOCONTENT: only the doc keys are checked, and it
-    # avoids returning the raw binary embedding that a decode_responses client cannot decode.
-    query_vec = struct.pack("<2f", 50.0, 100.0)
-    knn_result = await c_master.execute_command(
-        "FT.SEARCH",
-        "test_idx",
-        "*=>[KNN 2 @embedding $vec]",
-        "NOCONTENT",
-        "PARAMS",
-        "2",
-        "vec",
-        query_vec,
-    )
-    assert knn_result[0] >= 1
-    assert "item:50" in knn_result
-
-    # Start replication from new master to old replica
-    await start_replication(c_replica, master.port)
-
-    # Verify data replicated successfully
-    assert await c_replica.dbsize() == 100
-    assert await c_replica.hget("item:0", "name") == "Product 0"
-    assert await c_replica.hget("item:99", "name") == "Product 99"
-
-    # Verify KNN search works on old replica (index rebuilt from replicated data)
-    knn_result = await c_replica.execute_command(
-        "FT.SEARCH",
-        "test_idx",
-        "*=>[KNN 2 @embedding $vec]",
-        "NOCONTENT",
-        "PARAMS",
-        "2",
-        "vec",
-        query_vec,
-    )
-    assert knn_result[0] >= 1
-    assert "item:50" in knn_result
-
-
-async def test_replicating_mc_flags(df_factory):
-    master = df_factory.create(memcached_port=11211, proactor_threads=1)
-    replica = df_factory.create(
-        memcached_port=11212, proactor_threads=1, dbfilename=f"dump_{tmp_file_name()}"
-    )
-    df_factory.start_all([master, replica])
-
-    c_mc_master = pymemcache.Client(f"127.0.0.1:{master.mc_port}", default_noreply=False)
-
-    c_replica = replica.client()
-
-    assert c_mc_master.set("key1", "value0", noreply=True)
-    assert c_mc_master.set("key2", "value2", noreply=True, expire=3600, flags=123456)
-    assert c_mc_master.replace("key1", "value1", expire=4000, flags=2, noreply=True)
-
-    c_master = master.client()
-    for i in range(3, 100):
-        await c_master.set(f"key{i}", f"value{i}")
-
-    await start_replication(c_replica, master.port)
-
-    c_mc_replica = pymemcache.Client(f"127.0.0.1:{replica.mc_port}", default_noreply=False)
-
-    async def check_flag(key, flag):
-        res = c_mc_replica.raw_command("get " + key, "END\r\n").split()
-        # workaround sometimes memcached_client.raw_command returns empty str
-        if len(res) > 2:
-            assert res[2].decode() == str(flag)
-
-    await check_flag("key1", 2)
-    await check_flag("key2", 123456)
-
-    for i in range(1, 100):
-        assert c_mc_replica.get(f"key{i}") == str.encode(f"value{i}")
-
-
-async def test_double_take_over(df_factory, df_seeder_factory):
-    master = df_factory.create(proactor_threads=4, dbfilename="", admin_port=ADMIN_PORT)
-    replica = df_factory.create(proactor_threads=4, dbfilename="", admin_port=ADMIN_PORT + 1)
-    df_factory.start_all([master, replica])
-
-    seeder = df_seeder_factory.create(port=master.port, keys=1000, dbcount=5, stop_on_failure=False)
-    await seeder.run(target_deviation=0.1)
-
-    capture = await seeder.capture(port=master.port)
-
-    c_replica = replica.client()
-
-    logging.debug("start replication")
-    await start_replication(c_replica, master.admin_port)
-
-    logging.debug("running repltakover")
-    await c_replica.execute_command("REPLTAKEOVER 10")
-    assert await c_replica.execute_command("role") == master_role_reply([])
-
-    @assert_eventually
-    async def check_master_status():
-        assert master.proc.poll() == 0, "Master process did not exit correctly."
-
-    await check_master_status()
-
-    logging.debug("restart previous master")
-    master.start()
-    c_master = master.client()
-
-    logging.debug("start second replication")
-    await start_replication(c_master, replica.admin_port)
-
-    logging.debug("running second repltakover")
-    await c_master.execute_command("REPLTAKEOVER 10")
-    assert await c_master.execute_command("role") == master_role_reply([])
-
-    assert await seeder.compare(capture, port=master.port)
-
-
-async def test_replica_of_replica(df_factory):
-    # Can't connect a replica to a replica, but OK to connect 2 replicas to the same master
-    master = df_factory.create(proactor_threads=2)
-    replica = df_factory.create(proactor_threads=2)
-    replica2 = df_factory.create(proactor_threads=2)
-
-    df_factory.start_all([master, replica, replica2])
-
-    c_replica = replica.client()
-    c_replica2 = replica2.client()
-
-    assert await c_replica.execute_command(f"REPLICAOF localhost {master.port}") == "OK"
-
-    with pytest.raises(redis.exceptions.ResponseError):
-        await c_replica2.execute_command(f"REPLICAOF localhost {replica.port}")
-
-    assert await c_replica2.execute_command(f"REPLICAOF localhost {master.port}") == "OK"
-
-
-@pytest.mark.exclude_epoll
-@dfly_args({"proactor_threads": 1})
-async def test_memory_on_big_string_loading(df_factory):
-    """
-    In this test we want to make sure there is no spike in rss while loading big string value
-    1. insert 1 big value to master
-    2. replicate master
-    3. check rss peak memory on replica node
-    """
-    master = df_factory.create()
-    replica = df_factory.create()
-
-    df_factory.start_all([master, replica])
-    c_master = master.client()
-    c_replica = replica.client()
-
-    logging.debug("Populate with one big string")
-    await c_master.execute_command("DEBUG POPULATE 1 key 200000000 RAND")
-
-    async def get_memory(client, field):
-        info = await client.info("memory")
-        return info[field]
-
-    logging.debug("Start replication and wait for full sync")
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-    await wait_for_replicas_state(c_replica)
-
-    await c_replica.execute_command("memory decommit")
-    await asyncio.sleep(1)
-    replica_peak_memory = await get_memory(c_replica, "used_memory_peak_rss")
-    replica_used_memory = await get_memory(c_replica, "used_memory_rss")
-
-    logging.info(f"Replica Used memory {replica_used_memory}, peak memory {replica_peak_memory}")
-    assert replica_peak_memory < 1.1 * replica_used_memory
-
-    # Check replica data consistent
-    replica_data = await DebugPopulateSeeder.capture(c_replica)
-    master_data = await DebugPopulateSeeder.capture(c_master)
-    assert master_data == replica_data
-
-
-@pytest.mark.exclude_epoll
-@pytest.mark.parametrize(
-    "element_size, elements_number",
-    [(16, 30000), (30000, 16)],
-)
-@dfly_args({"proactor_threads": 1})
-async def test_big_containers(df_factory, element_size, elements_number):
-    master = df_factory.create()
-    replica = df_factory.create()
-
-    df_factory.start_all([master, replica])
-    c_master = master.client()
-    c_replica = replica.client()
-
-    logging.debug("Fill master with test data")
-    seeder = DebugPopulateSeeder(
-        key_target=50,
-        data_size=element_size * elements_number,
-        collection_size=elements_number,
-        variance=1,
-        samples=1,
-        types=["LIST", "SET", "ZSET", "HASH", "STREAM"],
-    )
-    await seeder.run(c_master)
-
-    async def get_memory(client, field):
-        info = await client.info("memory")
-        return info[field]
-
-    await asyncio.sleep(1)  # wait for heartbeat to update rss memory
-    used_memory = await get_memory(c_master, "used_memory_rss")
-
-    logging.debug("Start replication and wait for full sync")
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-    await wait_for_replicas_state(c_replica)
-
-    peak_memory = await get_memory(c_master, "used_memory_peak_rss")
-
-    logging.info(f"Used memory {used_memory}, peak memory {peak_memory}")
-    assert peak_memory < 1.1 * used_memory
-
-    await c_replica.execute_command("memory decommit")
-    await asyncio.sleep(1)
-    replica_peak_memory = await get_memory(c_replica, "used_memory_peak_rss")
-    replica_used_memory = await get_memory(c_replica, "used_memory_rss")
-
-    logging.info(f"Replica Used memory {replica_used_memory}, peak memory {replica_peak_memory}")
-    assert replica_peak_memory < 1.1 * replica_used_memory
-
-    # Check replica data consistent
-    replica_data = await DebugPopulateSeeder.capture(c_replica)
-    master_data = await DebugPopulateSeeder.capture(c_master)
-    assert master_data == replica_data
-
-
-async def test_master_too_big(df_factory):
-    master = df_factory.create(proactor_threads=4)
-    replica = df_factory.create(proactor_threads=2, maxmemory="600mb")
-
-    df_factory.start_all([master, replica])
-    c_master = master.client()
-    c_replica = replica.client()
-    await c_master.execute_command("DEBUG POPULATE 1000000 key 1000 RAND")
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-
-    # We should never sync due to used memory too high during full sync
-    with pytest.raises(TimeoutError):
-        await wait_available_async(c_replica, timeout=10)
-
-
-@dfly_args({"proactor_threads": 4})
-async def test_stream_approximate_trimming(df_factory):
-    master = df_factory.create()
-    replica = df_factory.create()
-
-    df_factory.start_all([master, replica])
-    c_master = master.client()
-    c_replica = replica.client()
-
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-    await wait_for_replicas_state(c_replica)
-
-    # Step 1: Populate master with 100 streams, each containing 200 entries
-    num_streams = 100
-    entries_per_stream = 200
-
-    for i in range(num_streams):
-        stream_name = f"stream{i}"
-        for j in range(entries_per_stream):
-            await c_master.execute_command("XADD", stream_name, "*", f"field{j}", f"value{j}")
-
-    # Step 2: Trim each stream to a random size between 70 and 200
-    for i in range(num_streams):
-        stream_name = f"stream{i}"
-        trim_size = random.randint(70, entries_per_stream)
-        await c_master.execute_command("XTRIM", stream_name, "MAXLEN", "~", trim_size)
-
-    # Wait for replica sync
-    await check_all_replicas_finished([c_replica], c_master)
-
-    # Check replica data consistent
-    master_data = await DebugPopulateSeeder.capture(c_master)
-    replica_data = await DebugPopulateSeeder.capture(c_replica)
-    assert master_data == replica_data
-
-    # Step 3: Trim all streams to 0
-    for i in range(num_streams):
-        stream_name = f"stream{i}"
-        await c_master.execute_command("XTRIM", stream_name, "MAXLEN", "0")
-
-    # Wait for replica sync
-    await check_all_replicas_finished([c_replica], c_master)
-
-    # Check replica data consistent
-    master_data = await DebugPopulateSeeder.capture(c_master)
-    replica_data = await DebugPopulateSeeder.capture(c_replica)
-    assert master_data == replica_data
-
-
-@dfly_args({"proactor_threads": 2})
-async def test_replicaof_does_not_flush_if_it_fails_to_connect(df_factory):
-    master = df_factory.create(proactor_threads=2)
-    replica = df_factory.create(proactor_threads=2)
-
-    df_factory.start_all([master, replica])
-    c_master = master.client()
-    c_replica = replica.client()
-
-    await c_master.execute_command("SET foo bar")
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-    await check_all_replicas_finished([c_replica], c_master)
-
-    res = await c_replica.execute_command("dbsize")
-    assert res == 1
-    with pytest.raises(redis.exceptions.ResponseError):
-        await c_replica.execute_command(f"REPLICAOF localhost {replica.port}")
-    res = await c_replica.execute_command("dbsize")
-    assert res == 1
-
-
-@dfly_args({"proactor_threads": 2})
-async def test_replicaof_inside_multi(df_factory):
-    master = df_factory.create()
-    replica = df_factory.create()
-    df_factory.start_all([master, replica])
-
-    async def replicate_inside_multi():
-        try:
-            c_master = master.client()
-            p = c_master.pipeline(transaction=True)
-            for i in range(5):
-                p.execute_command("dbsize")
-            p.execute_command(f"replicaof localhost {replica.port}")
-            await p.execute()
-            return True
-        except redis.exceptions.ResponseError:
-            return False
-
-    MULTI_COMMANDS_TO_ISSUE = 30
-    replication_commands = [
-        asyncio.create_task(replicate_inside_multi()) for _ in range(MULTI_COMMANDS_TO_ISSUE)
-    ]
-
-    num_successes = 0
-    for result in asyncio.as_completed(replication_commands, timeout=80):
-        num_successes += await result
-
-    logging.info(f"succeses: {num_successes}")
-    assert MULTI_COMMANDS_TO_ISSUE == num_successes
-
-
-@pytest.mark.large
-async def test_preempt_in_atomic_section_of_heartbeat(df_factory: DflyInstanceFactory):
-    master = df_factory.create(proactor_threads=1, serialization_max_chunk_size=100000000000)
-    replicas = [df_factory.create(proactor_threads=1) for i in range(2)]
-
-    # Start instances and connect clients
-    df_factory.start_all([master] + replicas)
-    c_master = master.client()
-    c_replicas = [replica.client() for replica in replicas]
-
-    total = 100000
-    await c_master.execute_command(f"DEBUG POPULATE {total} tmp 100 TYPE SET ELEMENTS 100")
-
-    thresehold = 50000
-    for i in range(thresehold):
-        rand = random.randint(1, 10)
-        await c_master.execute_command(f"EXPIRE tmp:{i} {rand} NX")
-
-    seeder = SeederV2(key_target=10_000)
-    fill_task = asyncio.create_task(seeder.run(master.client()))
-
-    for replica in c_replicas:
-        await replica.execute_command(f"REPLICAOF LOCALHOST {master.port}")
-
-    async with async_timeout.timeout(240):
-        await wait_for_replicas_state(*c_replicas)
-
-    await fill_task
-
-
-@pytest.mark.large
-async def test_bug_in_json_memory_tracking(df_factory: DflyInstanceFactory):
-    """
-    This test reproduces a bug in the JSON memory tracking.
-    """
-    random.seed(42)
-
-    master = df_factory.create(
-        proactor_threads=2,
-        serialization_max_chunk_size=1,
-        vmodule="replica=2,dflycmd=2,snapshot=1,rdb_save=1,rdb_load=1,journal_slice=2",
-    )
-    replicas = [df_factory.create(proactor_threads=2) for i in range(2)]
-
-    # Start instances and connect clients
-    df_factory.start_all([master] + replicas)
-    c_master = master.client()
-    c_replicas = [replica.client() for replica in replicas]
-
-    total = 100000
-    await c_master.execute_command(f"DEBUG POPULATE {total} tmp 1000 TYPE SET ELEMENTS 100")
-
-    threshold = 25000
-    for i in range(threshold):
-        rand = random.randint(1, 4)
-        await c_master.execute_command(f"EXPIRE tmp:{i} {rand} NX")
-
-    seeder = SeederV2(key_target=50_000)
-    fill_task = asyncio.create_task(seeder.run(master.client()))
-    await asyncio.sleep(0.2)
-
-    for replica in c_replicas:
-        await replica.execute_command(f"REPLICAOF LOCALHOST {master.port}")
-
-    async with async_timeout.timeout(240):
-        await wait_for_replicas_state(*c_replicas)
-
-    await seeder.stop(c_master)
-    await fill_task
-
-
-@pytest.mark.skip("fails, investigating")
-@pytest.mark.large
-@pytest.mark.opt_only
-@pytest.mark.parametrize("tagged_chunks", [True, False])
-@dfly_args({"proactor_threads": 2, "serialization_max_chunk_size": 5000, "compression_mode": "0"})
-async def test_big_huge_streaming_restart(df_factory: DflyInstanceFactory, tagged_chunks):
-    """
-    Restart replicating instance with huge values. Tests that interrupting the streaming process doesn't hinder retrying replication
-    """
-
-    master, replica = df_factory.create(
-        serialization_tagged_chunks=tagged_chunks
-    ), df_factory.create(proactor_threads=1)
-    df_factory.start_all([master, replica])
-    c_master, c_replica = master.client(), replica.client()
-
-    # Create huge values
-    await c_master.execute_command(
-        "debug", "populate", "2", "test", "1000", "rand", "type", "zset", "elements", "1000000"
-    )
-
-    done = False
-
-    async def ticker(c):
-        while not done:
-            await c_master.incr(f"key-{c}")
-
-    tickers = [asyncio.create_task(ticker(c)) for c in range(3)]
-
-    # Restart replication a few times
-    for _ in range(3):
-        print("Loop")
-        assert await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-        await asyncio.sleep(random.random() + 0.5)
-
-    # Wait for it to finish finally
-    async with async_timeout.timeout(75):
-        await wait_for_replicas_state(c_replica)
-
-    done = True
-    for t in tickers:
-        t.cancel()
-        try:
-            await t
-        except asyncio.CancelledError:
-            pass
-
-    await check_all_replicas_finished([c_replica], c_master)
-
-    # Check that everything is in sync
-    hashes = await asyncio.gather(*(SeederV2.capture(c) for c in [c_master, c_replica]))
-    assert len(set(hashes)) == 1
-
-    # No in-between errors occured
-    replica.stop()
-    lines = replica.find_in_logs("Duplicate zset fields detected")
-    assert len(lines) == 0
-
-
-@pytest.mark.large
-async def test_replica_snapshot_with_big_values_while_seeding(df_factory: DflyInstanceFactory):
-    proactors = 4
-    master = df_factory.create(proactor_threads=proactors, dbfilename="")
-    replica = df_factory.create(proactor_threads=proactors, dbfilename="")
-    df_factory.start_all([master, replica])
-    c_master = master.client()
-    c_replica = replica.client()
-
-    # 50% big values
-    seeder_config = dict(key_target=8_000, huge_value_target=4_000)
-    # Fill instance with test data
-    seeder = SeederV2(**seeder_config)
-    await seeder.run(c_master, target_deviation=0.01)
-
-    assert await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-    async with async_timeout.timeout(60):
-        await wait_for_replicas_state(c_replica)
-
-    # Start data stream
-    stream_task = asyncio.create_task(seeder.run(c_master))
-    await asyncio.sleep(1)
-
-    file_name = tmp_file_name()
-    assert await c_replica.execute_command(f"SAVE DF {file_name}") == "OK"
-    await seeder.stop(c_master)
-    await stream_task
-
-    await check_all_replicas_finished([c_replica], c_master)
-
-    # Check that everything is in sync
-    hashes = await asyncio.gather(*(SeederV2.capture(c) for c in [c_master, c_replica]))
-    assert len(set(hashes)) == 1
-
-    replica.stop()
-    lines = replica.find_in_logs("Exit SnapshotSerializer")
-    assert len(lines) == (proactors - 1)
-    for line in lines:
-        # We test the serializtion path of command execution
-        side_saved = extract_int_after_prefix("side saved ", line)
-        assert side_saved > 0
-
-    # Check that the produced rdb is loaded correctly
-    node = df_factory.create(dbfilename=file_name)
-    node.start()
-    c_node = node.client()
-    await wait_available_async(c_node)
-    assert await c_node.execute_command("dbsize") > 0
-    await c_node.execute_command("FLUSHALL")
-
-
-@pytest.mark.parametrize(
-    "use_takeover, backlog_len",
-    [(False, 2), (False, 1), (True, 1), (True, 10)],
-)
-async def test_partial_replication_on_same_source_master(df_factory, use_takeover, backlog_len):
-    master = df_factory.create()
-    replica1 = df_factory.create(shard_repl_backlog_len=backlog_len)
-    replica2 = df_factory.create()
-
-    df_factory.start_all([master, replica1, replica2])
-    c_master = master.client()
-    c_replica1 = replica1.client()
-    c_replica2 = replica2.client()
-
-    logging.debug("Fill master with test data")
-    seeder = DebugPopulateSeeder(key_target=50)
-    await seeder.run(c_master)
-
-    logging.debug("Start replication and wait for full sync")
-    await c_replica1.execute_command(f"REPLICAOF localhost {master.port}")
-    await wait_for_replicas_state(c_replica1)
-    await c_replica2.execute_command(f"REPLICAOF localhost {master.port}")
-    await wait_for_replicas_state(c_replica2)
-
-    # Send some traffic
-    seeder = SeederV2(key_target=8_000)
-    await seeder.run(c_master, target_deviation=0.01)
-
-    # Wait for all journal changes propagate to replicas
-    await check_all_replicas_finished([c_replica1, c_replica2], c_master)
-
-    if use_takeover:
-        # Promote first replica to master
-        await c_replica1.execute_command("REPLTAKEOVER 5")
-        if backlog_len > 1:
-            await c_replica1.execute_command("SET bar foo")
-            await c_replica1.execute_command("SET foo bar")
-
-    else:
-        # Promote first replica to master
-        await c_replica1.execute_command("REPLICAOF NO ONE")
-        await c_master.set("x", "y")
-        await c_master.set("x", "y")
-        await check_all_replicas_finished([c_replica2], c_master)
-
-    # Start replication with new master
-    await c_replica2.execute_command(f"REPLICAOF localhost {replica1.port}")
-
-    await check_all_replicas_finished([c_replica2], c_replica1)
-    # Validate data
-    if use_takeover:
-        hash1, hash2 = await asyncio.gather(
-            *(SeederV2.capture(c) for c in (c_replica1, c_replica2))
-        )
-        assert hash1 == hash2
-        s1 = await c_replica1.execute_command("dbsize")
-        s2 = await c_replica1.execute_command("dbsize")
-        assert s1 == s2
-
-    # Check we can takeover to the second replica
-    await c_replica2.execute_command("REPLTAKEOVER 5")
-
-    replica1.stop()
-    replica2.stop()
-    if use_takeover:
-        # Check logs for partial replication
-        lines = replica2.find_in_logs(f"Started partial sync with localhost:{replica1.port}")
-        assert len(lines) == 1
-        # Check no full sync logs
-        lines = replica2.find_in_logs(f"Started full sync with localhost:{replica1.port}")
-        assert len(lines) == 0
-    else:
-        lines = replica2.find_in_logs(f"Started full sync with localhost:{replica1.port}")
-        assert len(lines) == 1
-        # No partial sync after NO ONE
-        lines = replica2.find_in_logs(f"Started partial sync with localhost:{replica1.port}")
-        assert len(lines) == 0
-
-
-async def test_partial_replication_on_same_source_master_with_replica_lsn_inc(df_factory):
-    server1 = df_factory.create()
-    server2 = df_factory.create()
-    server3 = df_factory.create()
-    server4 = df_factory.create()
-
-    df_factory.start_all([server1, server2, server3, server4])
-    c_s2 = server2.client()
-    c_s3 = server3.client()
-    c_s4 = server4.client()
-
-    logging.debug("Start replication and wait for full sync")
-    await c_s2.execute_command(f"REPLICAOF localhost {server1.port}")
-    await wait_for_replicas_state(c_s2)
-    await c_s3.execute_command(f"REPLICAOF localhost {server1.port}")
-    await wait_for_replicas_state(c_s3)
-
-    # Promote server 2 to master
-    await c_s2.execute_command("REPLTAKEOVER 20")
-    # Make server 4 replica of server 2
-    await c_s4.execute_command(f"REPLICAOF localhost {server2.port}")
-    # Send some write command for lsn inc
-    # NOTE: using append temporarily because SET is omit-optimized with disables partial sync
-    for i in range(100):
-        await c_s2.append(i, "val")
-    # Make server 3 replica of server 2
-    await c_s3.execute_command(f"REPLICAOF localhost {server2.port}")
-
-    await check_all_replicas_finished([c_s3], c_s2)
-    await check_all_replicas_finished([c_s4], c_s2)
-
-    s2_sz = await c_s2.dbsize()
-    s3_sz = await c_s3.dbsize()
-    assert s2_sz == 100
-    assert s2_sz == s3_sz
-
-    s4_sz = await c_s4.dbsize()
-    assert s3_sz == s4_sz
-
-    server3.stop()
-    # Check logs for partial replication
-    lines = server3.find_in_logs(f"Started partial sync with localhost:{server2.port}")
-    assert len(lines) == 1
-
-
-async def test_replicate_hset_with_expiry(df_factory: DflyInstanceFactory):
-    master = df_factory.create(proactor_threads=2)
-    replica = df_factory.create(proactor_threads=2)
-
-    master.start()
-    replica.start()
-
-    cm = master.client()
-    await cm.execute_command("HSETEX key 86400 name 1234")
-
-    cr = replica.client()
-    await cr.execute_command(f"REPLICAOF localhost {master.port}")
-    await wait_available_async(cr)
-
-    result = await cr.hgetall("key")
-
-    assert "name" in result
-    assert result["name"] == "1234"
-
-
-async def test_bug_5221(df_factory):
-    master = df_factory.create(
-        proactor_threads=1,
-        cache_mode="true",
-        maxmemory="256mb",
-        enable_heartbeat_eviction="true",
-        eviction_memory_budget_threshold=0.9,
-    )
-    replica = df_factory.create(proactor_threads=4)
-    df_factory.start_all([master, replica])
-
-    c_master = master.client()
-    c_replica = replica.client()
-    await c_replica.execute_command(f"replicaof localhost {master.port}")
-
-    # Fill master with test data
-    seeder = SeederV2(key_target=22000, data_size=1000)
-    await seeder.run(c_master, target_deviation=0.01)
-    await asyncio.sleep(1)
-    await seeder.run(c_master, target_deviation=0.01)
-    res = await c_master.execute_command("dbsize")
-    assert res > 0
-
-
-@pytest.mark.parametrize("proactors", [1, 4, 6])
-@pytest.mark.parametrize("backlog_len", [1, 256, 1024, 1300])
-async def test_partial_sync(df_factory, proactors, backlog_len, proxy_factory):
-    keys = 5_000
-    if proactors > 1:
-        keys = 10_000
-
-    # We use lock_on_hashtag because we want to seed enough elements to one flow/journal such that
-    # the partial sync stales.
-    master = df_factory.create(
-        proactor_threads=proactors, shard_repl_backlog_len=backlog_len, lock_on_hashtags=True
-    )
-    replica = df_factory.create(proactor_threads=proactors)
-
-    df_factory.start_all([replica, master])
-
-    async def stream(client, total):
-        for i in range(0, total):
-            prefix = "{prefix}"
-            # Seed to one shard only. This will eventually cause one of the flows to become stale.
-            await client.execute_command(f"SET {prefix}foo{i} bar{i}")
-
-    async with replica.client() as c_replica, master.client() as c_master:
-        seeder = SeederV2(key_target=keys)
-        await seeder.run(c_master, target_deviation=0.01)
-
-        proxy = await proxy_factory(master.port)
-        await c_replica.execute_command(f"REPLICAOF localhost {proxy.port}")
-        # Reach stable sync
-        await wait_for_replicas_state(c_replica)
-        # Stream some elements
-        await stream(c_master, backlog_len)
-
-        proxy.drop_connection()
-        # Give time to detect dropped connection and reconnect
-        await asyncio.sleep(1.0)
-        # Partial synced here
-        await check_all_replicas_finished([c_replica], c_master)
-        hash1, hash2 = await asyncio.gather(*(SeederV2.capture(c) for c in (c_master, c_replica)))
-        assert hash1 == hash2
-
-        await proxy.close()
-        # Whoops we moved too much, no partial sync here
-        await stream(c_master, backlog_len + 10)
-        await proxy.start_serving()
-        await asyncio.sleep(1.0)
-
-        await check_all_replicas_finished([c_replica], c_master)
-
-        hash1, hash2 = await asyncio.gather(*(SeederV2.capture(c) for c in (c_master, c_replica)))
-        assert hash1 == hash2
-
-    master.stop()
-    replica.stop()
-    # Partial sync worked
-    lines = master.find_in_logs("Partial sync requested from LSN")
-    # Because we run with num_shards = proactors - 1
-    total_attempts = 1
-    if proactors > 1:
-        total_attempts = proactors - 1 + proactors - 2
-    assert len(lines) == total_attempts
-    # Second partial sync failed because of stale LSN
-    lines = master.find_in_logs("Partial sync requested from stale LSN")
-    assert len(lines) == 1
-
-
-async def test_mc_gat_replication(df_factory):
-    master = df_factory.create(memcached_port=11211, proactor_threads=1)
-    replica = df_factory.create(memcached_port=11212, proactor_threads=1)
-    df_factory.start_all([master, replica])
-
-    cm = pymemcache.Client(f"127.0.0.1:{master.mc_port}", default_noreply=False)
-
-    key = "foo"
-    value = b"bar"
-    not_found = b"NOTFOUND"
-    assert cm.set(key, value, noreply=True)
-
-    async with replica.client() as cl:
-        await cl.execute_command(f"REPLICAOF localhost {master.port}")
-        await wait_available_async(cl)
-
-    async def state_transitioned_stable(
-        init: bytes,
-        expected: bytes,
-        duration_sec=5,
-        sleep_sec=1,
-    ):
-        """
-        Asserts that the state goes from initial to expected and then stays at expected, observing state for duration_sec
-        """
-        _start = time.time()
-        transitioned = False
-        state = None
-        while time.time() - _start < duration_sec:
-            state = cr.get(key, not_found)
-            if not transitioned and state == expected:
-                transitioned = True
-            if transitioned:
-                assert (
-                    state == expected
-                ), f"state moved back to initial after transition {state=} {init=} {expected=}"
-            else:
-                assert state == init, f"unexpected state: {state=} {init=}"
-            await asyncio.sleep(sleep_sec)
-        return state == expected
-
-    cr = pymemcache.Client(f"127.0.0.1:{replica.mc_port}", default_noreply=False)
-
-    assert await state_transitioned_stable(not_found, value)
-
-    # Force the key to be removed by setting expiry in the past. Memcache treats expiry > 1 month as absolute from
-    # epoch, so 1 month + 1 second expires the key
-    month_plus_one = 60 * 60 * 24 * 30 + 1
-
-    # GAT|GATS are not directly exposed in the python client API
-    assert cm._fetch_cmd(b"gat", [str(month_plus_one), key], expect_cas=False) == {}
-
-    # The replica should eventually sync the delete operation
-    assert await state_transitioned_stable(value, not_found)
-
-    assert cm.set(key, value, noreply=True)
-    # expiry is set as now + 1000 seconds, which ensures the key will remain for the duration of the test
-    assert cm._fetch_cmd(b"gat", [str(1000), key], expect_cas=False) == {key: value}
-
-    # once the value is synced to the replica, assert that it remains stable and is not removed by setting expiry
-    assert await state_transitioned_stable(not_found, value)
-
-    result = cm._fetch_cmd(b"gats", [str(1000), key], expect_cas=True)
-    assert len(result) == 1 and key in result, f"missing expected key: {result=}"
-    expected_cas_ver = b"0"
-    assert result[key] == (value, expected_cas_ver), f"unexpected result for key: {result=}"
-
-
-@pytest.mark.skip("Fails constantly on CI")
-@pytest.mark.large
-@pytest.mark.parametrize("serialization_max_size", [1, 64000])
-async def test_replication_onmove_flow(df_factory, serialization_max_size):
-    master = df_factory.create(
-        proactor_threads=2,
-        cache_mode=True,
-        point_in_time_snapshot=False,
-        serialization_max_chunk_size=serialization_max_size,
-    )
-    replica = df_factory.create(proactor_threads=2)
-
-    df_factory.start_all([master, replica])
-    c_master = master.client()
-    c_replica = replica.client()
-
-    key_target = 100000
-    # Fill master with test data
-    await c_master.execute_command(f"DEBUG POPULATE {key_target} key 32 RAND TYPE hash ELEMENTS 10")
-    logging.debug("finished populate")
-
-    stop_event = asyncio.Event()
-
-    async def get_keys():
-        while not stop_event.is_set():
-            pipe = c_master.pipeline(transaction=False)
-            for _ in range(50):
-                id = random.randint(0, key_target)
-                pipe.hlen(f"key:{id}")
-            await pipe.execute()
-
-    get_task = asyncio.create_task(get_keys())
-    await asyncio.sleep(0.1)
-
-    # Start replication and wait for full sync
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-    await wait_for_replicas_state(c_replica)
-
-    info = await c_master.info("stats")
-    assert info["bump_ups"] >= 100
-
-    await check_all_replicas_finished([c_replica], c_master)
-    stop_event.set()
-    await get_task
-
-    # Check replica data consisten
-    hash1, hash2 = await asyncio.gather(*(SeederV2.capture(c) for c in (c_master, c_replica)))
-    assert hash1 == hash2
-
-    master.stop()
-    lines = master.find_in_logs("Exit SnapshotSerializer")
-    assert len(lines) > 0
-    for line in lines:
-        # We test the full sync on moved path execution
-        moved_saved = extract_int_after_prefix("moved_saved ", line)
-        logging.debug(f"Moved saves {moved_saved}")
-        assert moved_saved > 0
-
-
-@pytest.mark.large
-@dfly_args({"proactor_threads": 1})
-async def test_big_strings(df_factory):
-    master = df_factory.create(
-        proactor_threads=1, serialization_max_chunk_size=1, vmodule="snapshot=1"
-    )
-    replica = df_factory.create(proactor_threads=1)
-
-    df_factory.start_all([master, replica])
-    c_master = master.client()
-    c_replica = replica.client()
-
-    # 200kb
-    value_size = 200_000
-
-    async def get_memory(client, field):
-        info = await client.info("memory")
-        return info[field]
-
-    capacity = await get_memory(c_master, "prime_capacity")
-
-    seeder = DebugPopulateSeeder(
-        key_target=int(capacity * 0.7),
-        data_size=value_size,
-        collection_size=1,
-        variance=1,
-        samples=1,
-        types=["STRING"],
-    )
-    await seeder.run(c_master)
-
-    # sanity
-    capacity = await get_memory(c_master, "prime_capacity")
-    assert capacity < 8000
-
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-    await wait_for_replicas_state(c_replica)
-
-    # Check if replica data is consistent
-    replica_data = await DebugPopulateSeeder.capture(c_replica)
-    master_data = await DebugPopulateSeeder.capture(c_master)
-    assert master_data == replica_data
-
-    replica.stop()
-    master.stop()
-
-    lines = master.find_in_logs("Serialization peak bytes: ")
-    assert len(lines) == 1
-    # We test the serializtion path of command execution
-    line = lines[0]
-    peak_bytes = extract_int_after_prefix("Serialization peak bytes: ", line)
-    assert peak_bytes < value_size
-
-
-@pytest.mark.large
-async def test_takeover_bug_wrong_replica_checked_in_logs(df_factory):
-    master = df_factory.create(proactor_threads=4, vmodule="dflycmd=1")
-    replicas = [df_factory.create(proactor_threads=2) for _ in range(3)]
-    df_factory.start_all([master] + replicas)
-
-    c_master = master.client()
-    clients = [r.client() for r in replicas]
-
-    # Connect all replicas
-    for c in clients:
-        await c.execute_command(f"REPLICAOF localhost {master.port}")
-    await asyncio.gather(*[wait_available_async(c) for c in clients])
-
-    # Disconnect replica[1] to create lag
-    await clients[1].execute_command("REPLICAOF NO ONE")
-
-    # Write data that replica[1] will miss
-    pipe = c_master.pipeline()
-    for i in range(10000):
-        pipe.set(f"k{i}", "x" * 100)
-    await pipe.execute()
-
-    # Reconnect replica[1] and immediately takeover from replica[0]
-    await clients[1].execute_command(f"REPLICAOF localhost {master.port}")
-
-    await check_all_replicas_finished(clients, c_master)
-
-    await clients[0].execute_command("REPLTAKEOVER 10")
-
-    # Check master logs
-    master.stop(kill=False)
-
-    timeout_logs = master.find_in_logs(
-        f"Couldn't synchronize with replica for takeover in time: 127.0.0.1:{replicas[0].port}"
-    )
-    assert not timeout_logs
-
-
-@pytest.mark.large
-async def test_takeover_timeout_on_unresponsive_master(df_factory):
-    master = df_factory.create(proactor_threads=4)
-    replica = df_factory.create(proactor_threads=2)
-    df_factory.start_all([master, replica])
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    # Setup replication
-    await start_replication(c_replica, master.port)
-
-    # Write some data
-    for i in range(10):
-        await c_master.set(f"key{i}", f"val{i}")
-    await asyncio.sleep(0.2)
-
-    # PAUSE master process (SIGSTOP) - socket stays open but doesn't respondExpand commentComment on line R3629Code has comments. Press enter to view.
-    os.kill(master.proc.pid, signal.SIGSTOP)
-    logging.info(f"Paused master process {master.proc.pid}")
-
-    # Try takeover with 5 second timeout
-    # BUG: This will hang forever because SendNextPhaseRequest has no timeout
-    # FIXED: Should return error within ~15 seconds (5 + buffer)
-    start_time = time.time()
-    try:
-        await asyncio.wait_for(
-            c_replica.execute_command("REPLTAKEOVER 5"),
-            timeout=20,  # Should complete within 20 seconds
-        )
-        elapsed = time.time() - start_time
-        logging.info(f"Takeover completed in {elapsed:.1f}s")
-    except asyncio.TimeoutError:
-        elapsed = time.time() - start_time
-        pytest.fail(
-            f"BUG: REPLTAKEOVER hung for {elapsed:.1f}s without timeout. "
-            "SendNextPhaseRequest in replica.cc has no socket timeout."
-        )
-    except Exception as e:
-        # Expected: connection error or timeout error
-        elapsed = time.time() - start_time
-        logging.info(f"Takeover failed after {elapsed:.1f}s: {e}")
-        # Should fail quickly, not hang
-        assert elapsed < 20, f"Took too long: {elapsed:.1f}s"
-    finally:
-        # Resume master so it can be stopped properly
-        try:
-            os.kill(master.proc.pid, signal.SIGCONT)
-        except Exception:
-            pass
 
 
 async def test_replica_of_self(async_client):
@@ -3808,900 +1074,6 @@ async def test_repl_offset(df_factory):
     await with_timeout_link_down(c_replica3)
 
 
-async def test_partial_sync_with_different_shard_sizes(df_factory):
-    master = df_factory.create(proactor_threads=3)
-    replica1 = df_factory.create(proactor_threads=4)
-    replica2 = df_factory.create(proactor_threads=5)
-    replica3 = df_factory.create(proactor_threads=6)
-
-    df_factory.start_all([replica1, replica2, replica3, master])
-
-    c_replica1 = replica1.client()
-    c_replica2 = replica2.client()
-    c_replica3 = replica3.client()
-
-    c_master = master.client()
-
-    await c_master.execute_command("debug populate 5000")
-
-    await c_replica1.execute_command(f"replicaof localhost {master.port}")
-    await c_replica2.execute_command(f"replicaof localhost {master.port}")
-    await c_replica3.execute_command(f"replicaof localhost {master.port}")
-
-    seeder = SeederV2(key_target=100)
-    await seeder.run(c_master, target_deviation=0.01)
-
-    await check_all_replicas_finished([c_replica1, c_replica2, c_replica3], c_master)
-
-    await c_replica1.execute_command("repltakeover 5")
-    await c_replica2.execute_command(f"replicaof localhost {replica1.port}")
-    await c_replica3.execute_command(f"replicaof localhost {replica1.port}")
-
-    await check_all_replicas_finished([c_replica2, c_replica3], c_replica1)
-
-    for replica in (replica1, replica2, replica3):
-        replica.stop()
-
-    lines = replica2.find_in_logs(f"Started partial sync with localhost:{replica1.port}")
-    assert len(lines) == 0
-    lines = replica3.find_in_logs(f"Started partial sync with localhost:{replica1.port}")
-    assert len(lines) == 0
-
-
-@pytest.mark.large
-async def test_replica_reconnection_leaks_connections(df_factory: DflyInstanceFactory):
-    master = df_factory.create(proactor_threads=4)
-    replica = df_factory.create(proactor_threads=4)
-    df_factory.start_all([master, replica])
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    info = await c_master.info("clients")
-    baseline = info["connected_clients"]
-
-    num_cycles = 20
-    for _ in range(num_cycles):
-        await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-        await wait_for_replicas_state(c_replica)
-        await c_replica.execute_command("REPLICAOF NO ONE")
-
-    # Wait for connected_clients to stabilize (stop changing)
-    prev = None
-    async for info, breaker in info_tick_timer(c_master, "clients", timeout=10):
-        with breaker:
-            curr = info["connected_clients"]
-            assert curr == prev
-        prev = curr
-
-    leaked = prev - baseline
-    assert leaked == 0, f"connected_clients leaked {leaked} after {num_cycles} reconnect cycles"
-
-    await c_master.aclose()
-    await c_replica.aclose()
-
-
-@dfly_args({"proactor_threads": 2})
-@pytest.mark.replication
-async def test_xreadgroup_replication(replication):
-    _, _, c_master, [c_replica] = replication
-
-    async def compare_group_info(stream_key, expected_pending, expected_entries_read):
-        master_info = await c_master.execute_command(f"XINFO GROUPS {stream_key}")
-        replica_info = await c_replica.execute_command(f"XINFO GROUPS {stream_key}")
-
-        # Parse group info (format: [name, consumers, pending, last-delivered-id, entries-read, lag])
-        assert len(master_info) == len(replica_info)
-
-        for m_group, r_group in zip(master_info, replica_info):
-            m_dict = dict(zip(m_group[::2], m_group[1::2]))
-            r_dict = dict(zip(r_group[::2], r_group[1::2]))
-
-            assert m_dict["last-delivered-id"] == r_dict["last-delivered-id"]
-            assert m_dict["entries-read"] == r_dict["entries-read"]
-            assert m_dict["entries-read"] == expected_entries_read
-            assert m_dict["pending"] == r_dict["pending"]
-            assert m_dict["pending"] == expected_pending
-            assert m_dict["consumers"] == r_dict["consumers"]
-
-    # Case 1: Non-blocking path, NOACK
-    await c_master.execute_command("XGROUP CREATE mystream mygroup $ MKSTREAM")
-    await c_master.execute_command("XADD mystream * tmp tmp")
-    await c_master.execute_command("XREADGROUP GROUP mygroup worker1 NOACK STREAMS mystream >")
-
-    await check_all_replicas_finished([c_replica], c_master)
-    await compare_group_info("mystream", 0, 1)
-
-    # Case 2: Non-blocking path, with PEL
-    await c_master.execute_command("XADD mystream * tmp tmp")
-    await c_master.execute_command("XADD mystream * tmp tmp")
-    await c_master.execute_command("XREADGROUP GROUP mygroup worker1 STREAMS mystream >")
-
-    await check_all_replicas_finished([c_replica], c_master)
-    await compare_group_info("mystream", 2, 3)
-
-    # Case 3: Blocking path, NOACK
-
-    # Start blocking XREADGROUP in background
-    read_task = asyncio.create_task(
-        c_master.execute_command(
-            "XREADGROUP GROUP mygroup worker1 NOACK BLOCK 0 STREAMS mystream >"
-        )
-    )
-    # Let the blocking command start
-    await asyncio.sleep(0.1)
-    await c_master.execute_command("XADD mystream * tmp tmp")
-
-    await read_task
-
-    await check_all_replicas_finished([c_replica], c_master)
-    await compare_group_info("mystream", 2, 4)
-
-    # Case 4: Blocking path, with PEL
-
-    # Start blocking XREADGROUP in background
-    read_task = asyncio.create_task(
-        c_master.execute_command("XREADGROUP GROUP mygroup worker1 BLOCK 0 STREAMS mystream >")
-    )
-
-    await asyncio.sleep(0.1)
-    await c_master.execute_command("XADD mystream * tmp tmp")
-    await read_task
-
-    await check_all_replicas_finished([c_replica], c_master)
-    await compare_group_info("mystream", 3, 5)
-
-    await c_master.execute_command("flushall")
-    # Create consumer
-    await c_master.execute_command("XGROUP CREATE mystream mygroup $ MKSTREAM")
-    await c_master.execute_command("XADD mystream 2000-0 tmp tmp")
-    # Add to PEL but don't ack
-    await c_master.execute_command("XREADGROUP GROUP mygroup worker1 STREAMS mystream >")
-    await c_master.execute_command("XREADGROUP GROUP mygroup worker2 STREAMS mystream 2000-0")
-
-    await check_all_replicas_finished([c_replica], c_master)
-    await compare_group_info("mystream", 1, 1)
-
-
-"""
-Regression test for SIGSEGV when XREAD BLOCK unblocks with replication active.
-JournalXReadGroupIfNeeded accessed sitem.group->entries_read without checking that
-group is non-null (group is always null for XREAD, only set for XREADGROUP).
-"""
-
-
-@dfly_args({"proactor_threads": 2})
-async def test_xread_block_replication_crash_6975(df_factory):
-    master = df_factory.create()
-    replica = df_factory.create()
-
-    master.start()
-    replica.start()
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-    await wait_for_replicas_state(c_replica)
-
-    await c_master.execute_command("XADD mystream * tmp tmp")
-
-    # Start blocking XREAD - this will crash the server on unblock with replication active
-    # because JournalXReadGroupIfNeeded dereferences a null group pointer
-    read_task = asyncio.create_task(c_master.execute_command("XREAD BLOCK 0 STREAMS mystream $"))
-    await asyncio.sleep(0.1)
-    assert not read_task.done(), "XREAD BLOCK should still be blocking"
-
-    await c_master.execute_command("XADD mystream * field value")
-    result = await read_task
-
-    assert result is not None
-    await check_all_replicas_finished([c_replica], c_master)
-
-
-"""
-Test replication with mismatched dbnum between master and replica.
-"""
-
-
-@pytest.mark.parametrize("master_dbnum, replica_dbnum", [(8, 4), (4, 8)])
-@dfly_args({"proactor_threads": 2})
-async def test_replication_mismatched_dbnum(
-    df_factory: DflyInstanceFactory, master_dbnum, replica_dbnum
-):
-    """
-    Test replication with mismatched dbnum between master and replica.
-    When the replica has fewer DBs than the master, replication should still succeed for
-    the shared DBs (0..min-1). When the replica has more DBs, the extra DBs remain empty.
-    """
-    shared_dbs = min(master_dbnum, replica_dbnum)
-    master, [replica], c_master, [c_replica] = await setup_replication(
-        df_factory,
-        master_args={"dbnum": master_dbnum},
-        replica_args={"dbnum": replica_dbnum},
-        connect=False,
-    )
-
-    # Populate data only in shared DBs (within both master's and replica's range)
-    for db in range(shared_dbs):
-        c = master.client(db=db)
-        for i in range(50):
-            await c.set(f"key:{db}:{i}", f"val:{db}:{i}")
-        await c.close()
-
-    # Start replication
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-
-    async with async_timeout.timeout(10):
-        await wait_for_replicas_state(c_replica)
-
-    await check_all_replicas_finished([c_replica], c_master)
-
-    # Verify all data is present in the replica across shared DBs
-    for db in range(shared_dbs):
-        c_m = master.client(db=db)
-        c_r = replica.client(db=db)
-        for i in range(50):
-            assert await c_r.get(f"key:{db}:{i}") == await c_m.get(f"key:{db}:{i}")
-        await c_m.close()
-        await c_r.close()
-
-    # Verify extra DBs on replica are empty (when replica has more DBs than master)
-    if replica_dbnum > master_dbnum:
-        for db in range(master_dbnum, replica_dbnum):
-            c_r = replica.client(db=db)
-            assert await c_r.dbsize() == 0
-            await c_r.close()
-
-
-# BF.RESERVE with error_rate=0.00001 and capacity=1e9 creates a single bloom filter
-# of exactly 2^32 bytes (4 GiB). The chunked RDB loader used `unsigned` for the total
-# filter size, which silently overflowed to 0 and broke the RDB stream.
-# Verify that chunked replication of a large bloom filter works and the replicated filter
-# contains the expected items.
-@pytest.mark.large
-async def test_sbf_chunked_replication(df_factory: DflyInstanceFactory):
-    master = df_factory.create(
-        proactor_threads=1,
-        maxmemory="6G",
-    )
-    replica = df_factory.create(
-        proactor_threads=1,
-        maxmemory="6G",
-    )
-
-    df_factory.start_all([master, replica])
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    await c_master.execute_command("BF.RESERVE", "bf", "0.00001", "1000000000")
-    await c_master.execute_command("BF.ADD", "bf", "hello")
-
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-
-    async with async_timeout.timeout(240):
-        await wait_for_replicas_state(c_replica)
-
-    await check_all_replicas_finished([c_replica], c_master)
-    assert await c_replica.execute_command("BF.EXISTS", "bf", "hello") == 1
-
-
-# Verify chunked replication of a moderately-sized bloom filter (~1 GB) doesn't exceed
-# the max chunk size and correctly replicates all added items.
-@pytest.mark.large
-async def test_sbf_chunked_replication_chunk_size(df_factory: DflyInstanceFactory):
-    master = df_factory.create(
-        proactor_threads=1,
-        maxmemory="4G",
-    )
-    replica = df_factory.create(
-        proactor_threads=1,
-        maxmemory="4G",
-    )
-
-    df_factory.start_all([master, replica])
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    await c_master.execute_command("BF.RESERVE", "bf", "0.011", "400000000")
-
-    bf_size = await c_master.execute_command("MEMORY USAGE", "bf")
-    assert bf_size > 2**30, f"Bloom filter should be >1GB, got {bf_size}"
-
-    # Fix seed to have reproducible test results and log the seed for debugging
-    random_seed = random.getrandbits(64)
-    logging.info(f"Using random seed {random_seed} for test reproducibility")
-    test_rng = random.Random(random_seed)
-
-    num_items = 20000
-    random_items = [f"item:{i}" for i in test_rng.sample(range(1_000_000), num_items)]
-    await c_master.execute_command("BF.MADD", "bf", *random_items)
-
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-
-    async with async_timeout.timeout(240):
-        await wait_for_replicas_state(c_replica)
-
-    await check_all_replicas_finished([c_replica], c_master)
-
-    # Verify all added items exist on the replica
-    replica_results = await c_replica.execute_command("BF.MEXISTS", "bf", *random_items)
-    assert all(
-        replica_results
-    ), f"{replica_results.count(0)} of {len(random_items)} items missing on replica"
-
-    # Verify some non-added item doesn't exist on the replica
-    assert await c_replica.execute_command("BF.EXISTS", "bf", "not-added") == 0
-
-    master.stop()
-    replica.stop()
-
-    # We have set MAX_SBF_CHUNK_SIZE to 1 << 26. Double peak bytes size to be safe because of
-    # potential overhead that could be added.
-    MAX_SBF_CHUNK_SIZE = 2**27
-
-    lines = master.find_in_logs("Serialization peak bytes: ")
-    assert len(lines) == 1
-    line = lines[0]
-    peak_bytes = extract_int_after_prefix("Serialization peak bytes: ", line)
-    assert peak_bytes < MAX_SBF_CHUNK_SIZE
-
-
-# Per-sync timeouts share the test-level budget so individual waits don't
-# trip earlier than pytest.mark.timeout on slower CI runners.
-HNSW_LARGE_TEST_TIMEOUT_SEC = 900
-HNSW_LARGE_MARKS = [pytest.mark.large, pytest.mark.timeout(HNSW_LARGE_TEST_TIMEOUT_SEC)]
-
-
-def _hnsw_sync_timeout(num_docs: int) -> int:
-    return HNSW_LARGE_TEST_TIMEOUT_SEC if num_docs >= 100_000 else 120
-
-
-@pytest.mark.parametrize(
-    "master_threads, replica_threads, num_dims, num_docs",
-    [
-        pytest.param(3, 4, 4, 500, id="3t-4t-copied"),
-        pytest.param(4, 4, 4, 500, id="4t-4t-copied"),
-        pytest.param(4, 3, 4, 500, id="4t-3t-copied"),
-        pytest.param(4, 4, 256, 100, id="4t-4t-external"),
-        pytest.param(4, 4, 8, 100_000, id="4t-4t-100k", marks=HNSW_LARGE_MARKS),
-    ],
-)
-async def test_hnsw_search_replication_with_network_disruptions(
-    df_factory: DflyInstanceFactory,
-    master_threads: int,
-    replica_threads: int,
-    num_dims: int,
-    num_docs: int,
-    proxy_factory,
-):
-    """
-    Test HNSW search index replication under continuous traffic and a network disruption.
-
-    Creates a master with an HNSW vector index, starts concurrent write traffic and
-    search queries, replicates through a proxy, and drops the connection at a random
-    moment within the first 10 seconds (may hit full sync or stable sync).
-
-    When num_dims=256, vector data is 1024 bytes (>= default max_listpack_map_bytes),
-    forcing StringMap encoding and external HNSW vector pointers (copy_vector=false).
-    """
-    master = df_factory.create(proactor_threads=master_threads)
-    replica = df_factory.create(proactor_threads=replica_threads)
-    df_factory.start_all([master, replica])
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    seeder = HnswSearchSeeder(num_initial_docs=num_docs, num_dims=num_dims)
-    await seeder.create_index(c_master)
-    await seeder.seed_initial_docs(c_master)
-
-    proxy = await proxy_factory(master.port)
-    traffic_task = asyncio.create_task(seeder.run_traffic(c_master))
-    search_task = asyncio.create_task(seeder.run_search_queries(c_master))
-    await c_replica.execute_command(f"REPLICAOF localhost {proxy.port}")
-
-    sync_timeout = _hnsw_sync_timeout(num_docs)
-
-    # Wait for initial sync before running search queries on replica.
-    # During HNSW graph restoration with external vectors, nodes have
-    # uninitialized data pointers — search queries could dereference them.
-    await wait_available_async(c_replica, timeout=sync_timeout)
-    replica_search_task = asyncio.create_task(seeder.run_search_queries(c_replica))
-
-    try:
-        await asyncio.sleep(random.uniform(0, 10))
-        proxy.drop_connection()
-
-        # Give time to detect dropped connection and reconnect
-        await asyncio.sleep(1.0)
-
-        await wait_available_async(c_replica, timeout=sync_timeout)
-        seeder.stop()
-        await traffic_task
-        await search_task
-        await replica_search_task
-
-        # Log replica FT.INFO for debugging if assertion fails later
-        info = await c_replica.execute_command("FT.INFO", seeder.index_name)
-        logging.info(f"Replica FT.INFO: {info}")
-
-        await check_all_replicas_finished([c_replica], c_master, timeout=sync_timeout)
-        await seeder.verify(c_master, c_replica)
-
-    finally:
-        seeder.stop()
-        traffic_task.cancel()
-        search_task.cancel()
-        replica_search_task.cancel()
-
-
-@pytest.mark.parametrize(
-    "document_type, num_docs",
-    [
-        pytest.param("HASH", 300, id="HASH"),
-        pytest.param("JSON", 300, id="JSON"),
-        pytest.param("HASH", 100_000, id="HASH-100k", marks=HNSW_LARGE_MARKS),
-    ],
-)
-async def test_hnsw_failover_chain(
-    df_factory: DflyInstanceFactory, document_type: str, num_docs: int
-):
-    """
-    Primary → replica1 → REPLTAKEOVER → attach replica2 to promoted node.
-    The promoted node must still serve KNN, and a freshly attached replica
-    must rebuild the HNSW index from the promoted node's data.
-    """
-    master = df_factory.create(proactor_threads=2)
-    replica1 = df_factory.create(proactor_threads=2)
-    replica2 = df_factory.create(proactor_threads=2)
-    df_factory.start_all([master, replica1, replica2])
-
-    c_master = master.client()
-    c1 = replica1.client()
-    c2 = replica2.client()
-
-    sync_timeout = _hnsw_sync_timeout(num_docs)
-    seeder = HnswSearchSeeder(num_initial_docs=num_docs, num_dims=8, document_type=document_type)
-    await seeder.create_index(c_master)
-    await seeder.seed_initial_docs(c_master)
-
-    await c1.execute_command(f"REPLICAOF localhost {master.port}")
-    await wait_available_async(c1, timeout=sync_timeout)
-    await check_all_replicas_finished([c1], c_master, timeout=sync_timeout)
-    await seeder.verify(c_master, c1)
-
-    # Promote replica1. The master exits after REPLTAKEOVER completes.
-    await c1.execute_command("REPLTAKEOVER 5")
-    assert (await c1.execute_command("role"))[0] == "master"
-
-    # Attach replica2 to the promoted node and verify it rebuilds HNSW.
-    await c2.execute_command(f"REPLICAOF localhost {replica1.port}")
-    await wait_available_async(c2, timeout=sync_timeout)
-    await check_all_replicas_finished([c2], c1, timeout=sync_timeout)
-    await seeder.verify(c1, c2)
-
-
-@pytest.mark.parametrize("document_type", ["HASH", "JSON"])
-@pytest.mark.parametrize("data_type", ["INT8", "FLOAT16"])
-async def test_hnsw_search_replication_dtypes(
-    df_factory: DflyInstanceFactory, document_type: str, data_type: str
-):
-    """A non-float32 HNSW index replicates: the replica rebuilds it at the field's native
-    width and serves KNN identical to the master."""
-    master = df_factory.create(proactor_threads=2)
-    replica = df_factory.create(proactor_threads=2)
-    df_factory.start_all([master, replica])
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    seeder = HnswSearchSeeder(
-        num_initial_docs=200, num_dims=8, document_type=document_type, data_type=data_type
-    )
-    await seeder.create_index(c_master)
-    await seeder.seed_initial_docs(c_master)
-
-    await start_replication(c_replica, master.port)
-    await check_all_replicas_finished([c_replica], c_master)
-    await seeder.verify(c_master, c_replica)
-
-
-async def test_rm_replication(df_factory: DflyInstanceFactory):
-    """Test that RM command propagates deletions to replica and is rejected on replica."""
-    master = df_factory.create(proactor_threads=2)
-    replica = df_factory.create(proactor_threads=2)
-
-    master.start()
-    replica.start()
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    # Populate master with keys before replication starts
-    for i in range(20):
-        await c_master.set(f"key:{i}", f"val{i}")
-    for i in range(5):
-        await c_master.set(f"other:{i}", f"val{i}")
-
-    # Set up replication
-    await start_replication(c_replica, master.port)
-
-    # Verify replica has all keys
-    assert await c_replica.dbsize() == 25
-    logging.info("Replica has all keys")
-
-    # Run RM on master with a MATCH filter to delete only "key:*" keys
-    cursor = 0
-    while True:
-        result = await c_master.execute_command("RM", cursor, "MATCH", "key:*")
-        cursor = int(result[0])
-        if cursor == 0:
-            break
-
-    # Master should have only "other:*" keys left
-    assert await c_master.dbsize() == 5
-
-    # Wait for replication to propagate
-    await check_all_replicas_finished([c_replica], c_master)
-
-    # Replica should reflect deletions
-    assert await c_replica.dbsize() == 5
-    for i in range(5):
-        assert await c_replica.exists(f"other:{i}") == 1
-    for i in range(20):
-        assert await c_replica.exists(f"key:{i}") == 0
-
-    # RM must be rejected on replica (it's a write command)
-    with pytest.raises((aioredis.ResponseError, aioredis.ReadOnlyError)):
-        await c_replica.execute_command("RM", 0)
-
-
-@pytest.mark.parametrize(
-    "trigger_cmd, members, extra_setup",
-    [
-        # Originally from test_set_member_expiry_replication
-        (["SMEMBERS", "myset"], ["a", "b", "c"], []),
-        (["SUNION", "myset", "other"], ["a", "b", "c"], []),
-        (["SDIFF", "myset", "other"], ["a", "b", "c"], []),
-        (["SMISMEMBER", "myset", "a", "b", "c"], ["a", "b", "c"], []),
-        (["SSCAN", "myset", "0"], ["a", "b", "c"], []),
-        # Originally from test_set_member_expiry_replication_with_setup
-        (["SISMEMBER", "myset", "a"], ["a"], []),
-        (["SMOVE", "myset", "dst", "a"], ["a"], [["SADD", "dst", "x"]]),
-        (["SINTER", "myset", "other"], ["a", "b", "c"], [["SADD", "other", "a", "b", "c"]]),
-        (["FIELDEXPIRE", "myset", "100", "a"], ["a"], []),
-        (["FIELDTTL", "myset", "a"], ["a"], []),
-    ],
-    ids=[
-        "smembers",
-        "sunion",
-        "sdiff",
-        "smismember",
-        "sscan",
-        "sismember",
-        "smove",
-        "sinter",
-        "fieldexpire",
-        "fieldttl",
-    ],
-)
-async def test_set_member_expiry_replication(
-    df_factory: DflyInstanceFactory, trigger_cmd, members, extra_setup
-):
-    """
-    Verify that lazy set-member expiry on master is replicated to the replica.
-
-    When all members of a StringSet expire via lazy expiry (triggered by a read),
-    DeleteSetIfEmpty removes the key on master but does not journal the deletion.
-    This causes the replica to retain a stale key that no longer exists on master.
-
-    SISMEMBER/SMOVE are point lookups that only expire entries in the accessed bucket,
-    so a single member is used to guarantee the set becomes empty.
-    SINTER needs a second set to exist for the intersection code path to iterate.
-    """
-    master = df_factory.create(proactor_threads=2)
-    replica = df_factory.create(proactor_threads=2)
-
-    df_factory.start_all([master, replica])
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    # Set up replication before writing data to avoid race with TTL
-    await start_replication(c_replica, master.port)
-
-    await c_master.execute_command("SADDEX", "myset", "1", *members)
-    for cmd in extra_setup:
-        await c_master.execute_command(*cmd)
-
-    num_members = len(members)
-    assert await c_master.scard("myset") == num_members
-
-    await check_all_replicas_finished([c_replica], c_master)
-
-    assert await c_replica.scard("myset") == num_members
-
-    # Wait for members to expire
-    await asyncio.sleep(1)
-
-    # Trigger lazy expiry on master via a read command.
-    # Each variant exercises a different DeleteSetIfEmpty call site.
-    await c_master.execute_command(*trigger_cmd)
-
-    # The key should be gone on master
-    assert await c_master.exists("myset") == 0
-
-    await check_all_replicas_finished([c_replica], c_master)
-
-    assert await c_replica.exists("myset") == 0, (
-        f"Replica still has 'myset' after lazy expiry triggered by {trigger_cmd[0]}. "
-        "DeleteSetIfEmpty must journal the deletion so the replica stays in sync."
-    )
-
-
-@dfly_args({"proactor_threads": 4})
-@pytest.mark.timeout(60)
-async def test_replica_no_deadlock_on_disconnect(df_factory: DflyInstanceFactory):
-    """Replica must not deadlock when the master dies while a global command
-    (FLUSHALL) is being replicated across shards. The bug: one shard reads
-    the FLUSHALL from its socket buffer and enters Barrier::Wait inside
-    ExecuteTx, but CancelAllBlockingEntities already ran (triggered by
-    another shard whose socket broke first), so the new entry is never
-    cancelled and the shard fiber hangs forever."""
-    master = df_factory.create(proactor_threads=4)
-    replica = df_factory.create(proactor_threads=4)
-    df_factory.start_all([master, replica])
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    await c_master.execute_command("DEBUG", "POPULATE", "1000")
-    await c_replica.execute_command("REPLICAOF", "localhost", str(master.port))
-    await wait_available_async(c_replica)
-    await check_all_replicas_finished([c_replica], c_master)
-
-    # Stream Lua-based multi-shard traffic that generates global commands,
-    # then kill the master. The seeder's Lua scripts create complex multi-shard
-    # operations that widen the race window for the barrier deadlock.
-    stream_seeder = SeederV2(key_target=1000)
-    seed_task = asyncio.create_task(stream_seeder.run(c_master, target_deviation=0.1))
-    await asyncio.sleep(0.5)
-    await c_master.execute_command("FLUSHALL")
-    master.stop(kill=True)
-    seed_task.cancel()
-    try:
-        await seed_task
-    except (asyncio.CancelledError, Exception):
-        pass
-
-    # Restart master and re-point the replica.
-    master.start()
-    c_master = master.client()
-    await wait_available_async(c_master)
-    await c_master.execute_command("DEBUG", "POPULATE", "500")
-
-    # Re-point replica to the new master port. If the replica is deadlocked
-    # in an ExecuteTx barrier, even this command won't help it recover.
-    await c_replica.execute_command("REPLICAOF", "localhost", str(master.port))
-
-    await asyncio.wait_for(wait_for_replicas_state(c_replica), timeout=20)
-    await check_all_replicas_finished([c_replica], c_master, timeout=20)
-
-    assert await c_replica.dbsize() == await c_master.dbsize()
-    await c_replica.execute_command("REPLICAOF", "NO", "ONE")
-
-
-@dfly_args({"proactor_threads": 4})
-async def test_hnsw_external_vector_replication_crash(df_factory: DflyInstanceFactory):
-    """
-    Minimal reproducer for SIGSEGV during HNSW replication with external vectors.
-
-    The HNSW graph is global (shared across shards) but is_restoring_vectors_ is per-shard.
-    When one shard finishes RestoreGlobalVectorIndices and starts accepting new HSETs,
-    addPoint() traverses the global graph and may dereference nullptr data pointers
-    from nodes belonging to shards that haven't finished restoration yet.
-    """
-    master = df_factory.create(proactor_threads=4)
-    replica = df_factory.create(proactor_threads=4)
-    df_factory.start_all([master, replica])
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    # dim=256 forces external vectors (copy_vector=false).
-    # 500 docs creates a larger HNSW graph, widening the restoration window.
-    seeder = HnswSearchSeeder(num_initial_docs=500, num_dims=256)
-    await seeder.create_index(c_master)
-    await seeder.seed_initial_docs(c_master)
-
-    # Start heavy traffic on master BEFORE replication begins.
-    # This ensures journal events (HSETs) arrive on the replica while HNSW graph
-    # is being restored, triggering addPoint() with nullptr data pointers.
-    traffic_task = asyncio.create_task(seeder.run_traffic(c_master, sleep_interval=0.001))
-
-    await start_replication(c_replica, master.port)
-
-    seeder.stop()
-    await traffic_task
-
-    await check_all_replicas_finished([c_replica], c_master, timeout=60)
-
-
-async def _ft_num_docs(client, index_name: str) -> int:
-    info = await client.execute_command("FT.INFO", index_name)
-    info_map = dict(zip(info[::2], info[1::2]))
-    return int(info_map["num_docs"])
-
-
-@dfly_args({"proactor_threads": 4})
-@pytest.mark.parametrize(
-    "docs_keep",
-    [
-        pytest.param(2000, id="small"),
-        pytest.param(100_000, id="100k", marks=HNSW_LARGE_MARKS),
-    ],
-)
-async def test_hnsw_multi_replica_with_concurrent_index_ops(
-    df_factory: DflyInstanceFactory, docs_keep: int
-):
-    """
-    Fan-out HNSW replication with a pre-existing index dropped and a new index
-    created shortly after the replicas connect. Verifies doc-count parity and
-    KNN coverage across master and every replica.
-
-    Covers next:
-      - Multiple replicas (3 fan-out)
-      - FT.DROPINDEX of a pre-existing index, 0.5-5s after replicas connect
-      - FT.CREATE of a new index, 0.5-5s after replicas connect
-      - Document-count parity master/replica per surviving index
-      - 100k-document HNSW main index (large variant)
-    """
-    master = df_factory.create()
-    replicas = [df_factory.create(proactor_threads=2) for _ in range(3)]
-    df_factory.start_all([master] + replicas)
-
-    c_master = master.client()
-    c_replicas = [r.client() for r in replicas]
-
-    seeder_keep = HnswSearchSeeder(
-        num_initial_docs=docs_keep, num_dims=8, index_name="idx_keep", prefix="keep:"
-    )
-    seeder_drop = HnswSearchSeeder(
-        num_initial_docs=1000, num_dims=8, index_name="idx_drop", prefix="drop:"
-    )
-    seeder_new = HnswSearchSeeder(
-        num_initial_docs=500, num_dims=8, index_name="idx_new", prefix="new:"
-    )
-
-    await seeder_keep.create_index(c_master)
-    await seeder_keep.seed_initial_docs(c_master)
-    await seeder_drop.create_index(c_master)
-    await seeder_drop.seed_initial_docs(c_master)
-
-    traffic_task = asyncio.create_task(seeder_keep.run_traffic(c_master))
-    sync_timeout = _hnsw_sync_timeout(docs_keep)
-
-    try:
-        await asyncio.gather(
-            *(c.execute_command(f"REPLICAOF localhost {master.port}") for c in c_replicas)
-        )
-
-        # Drop the pre-existing index and create the new one, each at an
-        # independent random offset in the (0.5, 5)s window after REPLICAOF.
-        # The randomness is intentional: it spreads the ops across the replica's
-        # full-sync window to exercise the indexing race. Log the offsets so a
-        # failure is reproducible.
-        drop_delay = random.uniform(0.5, 5)
-        create_delay = random.uniform(0.5, 5)
-        logging.info(f"drop_delay={drop_delay:.2f}s create_delay={create_delay:.2f}s")
-
-        async def drop_after_delay():
-            await asyncio.sleep(drop_delay)
-            await c_master.execute_command("FT.DROPINDEX", "idx_drop")
-
-        async def create_after_delay():
-            await asyncio.sleep(create_delay)
-            await seeder_new.create_index(c_master)
-            await seeder_new.seed_initial_docs(c_master)
-
-        await asyncio.gather(drop_after_delay(), create_after_delay())
-
-        master_indices = set(await c_master.execute_command("FT._LIST"))
-        assert master_indices == {
-            "idx_keep",
-            "idx_new",
-        }, f"Master indices unexpected: {master_indices}"
-
-        await wait_available_async(c_replicas, timeout=sync_timeout)
-        seeder_keep.stop()
-        await traffic_task
-        await check_all_replicas_finished(c_replicas, c_master, timeout=sync_timeout)
-
-        for c in c_replicas:
-            idx_list = set(await c.execute_command("FT._LIST"))
-            assert idx_list == {"idx_keep", "idx_new"}, f"Unexpected indices: {idx_list}"
-
-        # HNSW background indexing on the replica can lag the journal LSN, so
-        # poll until num_docs converges. Fan out across all (replica × index) pairs.
-        @assert_eventually(timeout=sync_timeout)
-        async def _converged(client, index_name: str, expected: int):
-            actual = await _ft_num_docs(client, index_name)
-            assert actual == expected, f"{index_name}: replica={actual} master={expected}"
-
-        index_to_expected = {
-            name: await _ft_num_docs(c_master, name) for name in ("idx_keep", "idx_new")
-        }
-        await asyncio.gather(
-            *(
-                _converged(c, name, expected)
-                for c in c_replicas
-                for name, expected in index_to_expected.items()
-            )
-        )
-
-        await asyncio.gather(*(seeder_keep.verify(c_master, c) for c in c_replicas))
-        await asyncio.gather(*(seeder_new.verify(c_master, c) for c in c_replicas))
-
-    finally:
-        seeder_keep.stop()
-        traffic_task.cancel()
-        try:
-            await traffic_task
-        except (asyncio.CancelledError, Exception):
-            pass
-
-
-async def test_snapshot_load_replication(df_factory: DflyInstanceFactory):
-    dbfilename = f"dump_{tmp_file_name()}"
-
-    master = df_factory.create()
-    replica = df_factory.create()
-    df_factory.start_all([master, replica])
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    # Populate initial data and save a snapshot.
-    seeder = DebugPopulateSeeder(key_target=1000, data_size=100)
-    await seeder.run(c_master)
-    await c_master.execute_command("SAVE", "DF", dbfilename)
-    await c_master.execute_command("FLUSHALL")
-
-    await c_replica.execute_command("REPLICAOF", "localhost", str(master.port))
-    await wait_available_async(c_replica)
-
-    # Stream writes during DFLY LOAD to exercise the race between journal
-    # writes and the load that bypasses the journal. LOADING state rejects
-    # seeder Lua scripts, so the seeder task may fail.
-    stream_seeder = SeederV2(key_target=500)
-    seed_task = asyncio.create_task(stream_seeder.run(c_master, target_deviation=0.1))
-    await asyncio.sleep(
-        0.5
-    )  # Let the seeder start and write some data before we load the snapshot.
-
-    await c_master.execute_command("DFLY", "LOAD", f"{dbfilename}-summary.dfs")
-
-    await asyncio.sleep(0.5)  # Let the seeder fail because of the loading state.
-    seed_task.cancel()
-    try:
-        await seed_task
-    except (asyncio.CancelledError, Exception):
-        pass
-
-    # Wait for the replica to complete the new full sync.
-    await wait_for_replicas_state(c_replica)
-    await check_all_replicas_finished([c_replica], c_master)
-
-    master_capture = await DebugPopulateSeeder.capture(c_master)
-    replica_capture = await DebugPopulateSeeder.capture(c_replica)
-    assert master_capture == replica_capture
-
-    await c_replica.execute_command("REPLICAOF", "NO", "ONE")
-
-
 async def test_client_list_replication_types(df_factory: DflyInstanceFactory):
     master = df_factory.create(proactor_threads=2)
     replica = df_factory.create(proactor_threads=1)
@@ -4763,171 +1135,6 @@ async def test_client_list_replication_types(df_factory: DflyInstanceFactory):
         assert parse_client_list(await c_replica.execute_command("CLIENT LIST TYPE master")) == []
 
     await link_gone()
-
-
-async def test_v134_replica_crashes_v138_master_on_reconnect(df_factory: DflyInstanceFactory):
-    dfly_version = "v1.34.2"
-    released_dfly_path = download_dragonfly_release(dfly_version)
-
-    old_master = df_factory.create(path=released_dfly_path, version=1.34, proactor_threads=2)
-    new_master = df_factory.create(proactor_threads=2)
-    old_replica = df_factory.create(path=released_dfly_path, version=1.34, proactor_threads=2)
-    df_factory.start_all([old_master, new_master, old_replica])
-
-    c_old_master = old_master.client()
-    c_new_master = new_master.client()
-    c_old_replica = old_replica.client()
-
-    await c_old_master.set("key", "value")
-
-    # Both new_master and old_replica replicate from old_master and reach stable sync.
-    await start_replication(c_new_master, old_master.port)
-    await start_replication(c_old_replica, old_master.port)
-
-    await c_new_master.execute_command("REPLTAKEOVER 5")
-
-    await c_old_replica.execute_command(f"REPLICAOF localhost {new_master.port}")
-
-    @assert_eventually(timeout=10)
-    async def replica_connected():
-        info = await c_new_master.info("replication")
-        assert info.get("connected_slaves", 0) == 1
-
-    await replica_connected()
-
-    # Trigger cleanup: disconnect old_replica so new_master runs ~SliceSnapshot.
-    await c_old_replica.execute_command("REPLICAOF NO ONE")
-
-    @assert_eventually(timeout=5)
-    async def new_master_alive():
-        assert new_master.proc.poll() is None, (
-            f"new_master (v1.38+) crashed (exit={new_master.proc.poll()}) when a v1.34 replica "
-            "reconnected after promotion — regression of the v1.34→v1.38 rolling-upgrade crash."
-        )
-
-    await new_master_alive()
-
-
-@pytest.mark.asyncio
-async def test_bgsave_during_stable_sync(df_factory: DflyInstanceFactory):
-    """
-    shard_stable_sync_read when a BGSAVE is running on the replica concurrently
-    """
-    master, [replica], c_master, [c_replica] = await setup_replication(
-        df_factory,
-        master_args={"proactor_threads": 8, "num_shards": 8},
-        replica_args={
-            "proactor_threads": 8,
-            "num_shards": 8,
-            "dir": "{DRAGONFLY_TMP}/",
-            "dbfilename": "test-replica-bgsave",
-        },
-        connect=False,
-    )
-    del_clients = [master.client() for _ in range(16)]
-
-    await c_master.execute_command("DEBUG", "POPULATE", "100000", "k", "100")
-
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-    await wait_for_replicas_state(c_replica)
-
-    stop_event = asyncio.Event()
-
-    async def del_loop(client, slot):
-        i = slot
-        while not stop_event.is_set():
-            try:
-                pipe = client.pipeline(transaction=False)
-                for j in range(50):
-                    pipe.delete(f"k:{(i + j) % 100000}")
-                    pipe.set(f"k:{(i + j) % 100000}", f"v{i + j}")
-                await pipe.execute()
-                i += 50
-            except Exception:
-                stop_event.set()
-                break
-
-    async def bgsave_loop():
-        for _ in range(20):
-            if stop_event.is_set():
-                break
-            try:
-                await c_replica.execute_command("BGSAVE")
-                while await is_saving(c_replica):
-                    await asyncio.sleep(0)
-            except Exception:
-                stop_event.set()
-                break
-        stop_event.set()
-
-    await asyncio.gather(
-        bgsave_loop(),
-        *[del_loop(c, i * 6250) for i, c in enumerate(del_clients)],
-        return_exceptions=True,
-    )
-
-    for c in del_clients:
-        await c.aclose()
-
-    await c_master.aclose()
-    await c_replica.aclose()
-    master.stop()
-    replica.stop()
-
-    bad_lines = replica.find_in_logs("Unexpected fiber")
-    assert not bad_lines, (
-        "DbSlice change callback fired from shard_stable_sync_read during BGSAVE "
-        "(SerializerBase::OnChangeBlocking invariant violated).\n" + "\n".join(bad_lines)
-    )
-
-
-# Verify chunked CF replication stays within the max chunk size budget.
-@pytest.mark.large
-async def test_cf_chunked_replication_chunk_size(df_factory: DflyInstanceFactory):
-    master = df_factory.create(proactor_threads=1, maxmemory="512m")
-    replica = df_factory.create(proactor_threads=1, maxmemory="512m")
-
-    df_factory.start_all([master, replica])
-
-    c_master = master.client()
-    c_replica = replica.client()
-
-    await c_master.execute_command("CF.RESERVE", "cf", "100000000")
-
-    cf_size = await c_master.execute_command("MEMORY USAGE", "cf")
-    assert cf_size > 2**26, f"Cuckoo filter should be >64MB, got {cf_size}"
-
-    random_seed = random.getrandbits(64)
-    logging.info(f"Using random seed {random_seed} for test reproducibility")
-    test_rng = random.Random(random_seed)
-
-    num_items = 200
-    random_items = [f"item:{i}" for i in test_rng.sample(range(1_000_000), num_items)]
-    async with c_master.pipeline(transaction=False) as pipe:
-        for item in random_items:
-            pipe.execute_command("CF.ADD", "cf", item)
-        await pipe.execute()
-
-    await c_replica.execute_command(f"REPLICAOF localhost {master.port}")
-
-    async with async_timeout.timeout(240):
-        await wait_for_replicas_state(c_replica)
-
-    replica_results = await asyncio.gather(
-        *(c_replica.execute_command("CF.EXISTS", "cf", item) for item in random_items)
-    )
-    missing = sum(1 for r in replica_results if not r)
-    assert missing == 0, f"{missing} of {num_items} items missing on replica"
-
-    master.stop()
-    replica.stop()
-
-    # kFilterChunkSize = 1 << 26 (64MB); peak should stay under 2x that.
-    max_cf_chunk_size = 2**27
-    lines = master.find_in_logs("Serialization peak bytes: ")
-    assert len(lines) == 1
-    peak_bytes = extract_int_after_prefix("Serialization peak bytes: ", lines[0])
-    assert peak_bytes < max_cf_chunk_size
 
 
 @dfly_args({"proactor_threads": 2})
